% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{add_quadrant_labels}
\alias{add_quadrant_labels}
\title{Add quadrant labels to a ggplot representing a scatterplot}
\usage{
add_quadrant_labels(plot, x, y)
}
\arguments{
\item{plot}{ggplot object representing a scatterplot}

\item{x}{horizontal y coordinate to separate the quadrants}

\item{y}{vertical x coordinate to separate the quadrants}
}
\value{
ggplot object with quadrant boundaries and text labels added
}
\description{
This function takes in a ggplot that is representing a scatterplot, a horizontal y coordinate, and a vertical x coordinate, and adds quadrant boundaries and text labels to the plot to indicate how many points are in each section of the scatterplot.
}
