% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMembrane_based.R
\name{autoProcessRawSerObjs.prime}
\alias{autoProcessRawSerObjs.prime}
\title{autoProcessRawSerObjs.prime: An automated and basic pipeline to process raw Seurat objects obtained form prime-seq with CellMembrane package.}
\usage{
autoProcessRawSerObjs.prime(
  seuratObjects,
  verbose = T,
  doDoubletFinder = T,
  dropDoublets = F,
  doAppendCellHashing = T,
  featureData = NULL,
  doAppendCITEseq = T,
  doRawCountFilter = T,
  doMergeObjects = F,
  mergeObjectName = "MergedObjs",
  doNomalizeAndScale = T,
  doRegressBasic = T,
  doRegresspmito = T,
  doRemoveCellCycle = F,
  doDimReduxTrio = T,
  doSingleR = T,
  nCount_RNA.low = 0,
  nFeature.low = 200,
  pMito.low = 0,
  nCount_RNA.high = 20000,
  nFeature.high = 5000,
  pMito.high = 0.15,
  variableFeatureSelectionMethod = "vst",
  nVariableFeatures = 2000,
  mean.cutoff = c(0.0125, 3),
  dispersion.cutoff = c(0.5, Inf),
  block.size = 1000,
  variableGenesInclusion = NULL,
  variableGenesExclusion = NULL,
  npcs = 50,
  minDimsToUse = 15
)
}
\arguments{
\item{seuratObjects, }{A named list of Seurat object.}

\item{verbose, }{set T/F for print to screen msgs}

\item{doDoubletFinder, }{removed doublets from the seurat objects}

\item{doAppendCellHashing, }{filter cells based on defined parameters}

\item{doAppendCITEseq, }{filter cells based on defined parameters}

\item{doRawCountFilter, }{filter cells based on defined parameters}

\item{doMergeObjects, }{filter cells based on defined parameters}

\item{doNomalizeAndScale, }{filter cells based on defined parameters}

\item{doRegressBasic, }{filter cells based on defined parameters}

\item{doRegresspmito, }{filter cells based on defined parameters}

\item{doRemoveCellCycle, }{filter cells based on defined parameters}

\item{doDimReduxTrio, }{filter cells based on defined parameters}

\item{doSingleR, }{filter cells based on defined parameters}

\item{nCount_RNA.low}{If doCellFilter=T, cells with nCount_RNA below this value will be filtered}

\item{nFeature.low}{If doCellFilter=T, cells with nFeature below this value will be filtered}

\item{pMito.low}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{nCount_RNA.high}{If doCellFilter=T, cells with nCount_RNA above this value will be filtered}

\item{nFeature.high}{If doCellFilter=T, cells with nFeature above this value will be filtered}

\item{pMito.high}{If doCellFilter=T, cells with percent mito above this value will be filtered}

\item{variableFeatureSelectionMethod}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{nVariableFeatures}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{mean.cutoff}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{dispersion.cutoff}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{block.size}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{variableGenesInclusion}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{variableGenesExclusion}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{npcs}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{minDimsToUse}{If doCellFilter=T, cells with percent mito  below this value will be filtered}
}
\value{
A processed seruat object
}
\description{
This is the entry point for processing raw scRNAseq data with Seurat & CellMembrane.
}
