% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{barplot_vectors}
\alias{barplot_vectors}
\title{barplot vectors side by side}
\usage{
barplot_vectors(vec1, vec2, title = "", QuantThr = 0.66)
}
\arguments{
\item{vec1}{a named vector}

\item{vec2}{a secibd named vector}

\item{title}{title}

\item{QuantThr}{color pink via quantiles}
}
\value{
If returnDF is `TRUE`, the plotted data frame is returned.
}
\description{
barplot vectors side by side
}
