% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{BetterViolins}
\alias{BetterViolins}
\title{BetterViolins}
\usage{
BetterViolins(
  SerObj = NULL,
  feature2plot = "nCount_RNA",
  featuredummy = "nFeature_RNA",
  group.by = "ExpID",
  downsampleN = 500,
  title = "",
  ComplexPlot = F,
  log10Y = T
)
}
\arguments{
\item{SerObj}{a data frame or a Seurat object containing the expression data}

\item{feature2plot}{a string indicating the feature to plot on the y-axis (default: "nCount_RNA")}

\item{featuredummy}{a string indicating a dummy feature to plot on the x-axis (default: "nFeature_RNA")}

\item{group.by}{a string indicating the grouping variable (default: "ExpID")}

\item{downsampleN}{an integer indicating the number of cells to downsample to (default: 500)}

\item{title}{a string indicating the title of the plot (default: "")}

\item{ComplexPlot}{a boolean indicating whether to generate a complex plot or not (default: FALSE)}

\item{log10Y}{a boolean indicating whether to use log10 transformation on the y-axis (default: TRUE)}
}
\value{
a ggplot object
}
\description{
This function generates violin plots with customizable features.
}
