% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Bool.CellsGT}
\alias{Bool.CellsGT}
\title{Bool.CellsGT}
\usage{
Bool.CellsGT(SerObj, thresh, layer = "data", assay = "RNA", gene = NULL)
}
\arguments{
\item{SerObj}{A Seurat object.}

\item{thresh}{A numeric threshold value.}

\item{layer}{The layer name of the Seurat object to retrieve the data from.}

\item{assay}{The assay name to retrieve the data from.}

\item{gene}{The name of the gene to subset the data.}
}
\description{
This function checks whether the values in a specific gene's assay data of a Seurat object are greater than a given threshold.
}
