% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{CalculateModuleScoreAvg}
\alias{CalculateModuleScoreAvg}
\title{CalculateModuleScoreAvg}
\usage{
CalculateModuleScoreAvg(
  seuratObj,
  genes.list = NULL,
  genes.pool = NULL,
  n.bin = 25,
  ctrl.size = 100,
  assay = NULL,
  seed.use = 1
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{genes.list}{A character vector specifying the genes to calculate the module scores.}

\item{genes.pool}{A character vector specifying the genes to use for calculating the average expression.}

\item{n.bin}{An integer specifying the number of bins for partitioning the average expression values.}

\item{ctrl.size}{An integer specifying the size of the control group for calculating module scores.}

\item{assay}{The assay name to retrieve the data from. If NULL, the default assay of the Seurat object is used.}

\item{seed.use}{An integer specifying the seed value for random number generation.}
}
\description{
This function calculates the module scores for each cell in a Seurat object based on the average expression of a given gene list.
}
