% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{CellScore_FeaturePlot}
\alias{CellScore_FeaturePlot}
\title{Cell Score Feature Plot}
\usage{
CellScore_FeaturePlot(
  SerObj,
  reduction = "inmf",
  compN = 1,
  plot = T,
  plotting_reduction = "umap",
  raster = F,
  raster.dpi = c(2000, 2000),
  pt.size = 0.1,
  order = T
)
}
\arguments{
\item{SerObj}{A Seurat SerObj.}

\item{reduction}{which reduction to use to get cell scores}

\item{plot}{if true returns plot else returns SerObj with CellScore}

\item{plotting_reduction}{a reduction for plotting.}

\item{raster}{raster or not}

\item{raster.dpi}{default c(2000, 2000)}

\item{pt.size}{point size passed to featureplot}

\item{order}{default T passed to featureplot}
}
\value{
plot or SerObj
}
\description{
Plots a feature plot of cell scores from a specific reduction
}
\details{
General principles behind it:
}
\examples{
CellScore_FeaturePlot(SerObj = SerObj, compN = 3, reduction = "inmf",
plotting_reduction = "liger_r1_umap", plot = T,
pt.size = .5)

}
