% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{CellScore_RidgePlot}
\alias{CellScore_RidgePlot}
\title{Cell Score Ridge Plot}
\usage{
CellScore_RidgePlot(
  SerObj,
  reduction = "inmf",
  group.by = "orig.ident",
  compN = 1,
  plot = T,
  cols = NULL
)
}
\arguments{
\item{SerObj}{A Seurat SerObj.}

\item{reduction}{which reduction to use to get cell scores}

\item{plot}{if true returns plot else returns SerObj with CellScore}

\item{cols}{color vector passed to VlnPlot}

\item{raster}{raster or not}

\item{pt.size}{point size passed to featureplot}
}
\value{
plot or SerObj
}
\description{
Plots a ridge plot of cell scores from a specific reduction
}
\details{
General principles behind it:
}
\examples{
CellScore_RidgePlot(SerObj = ComboSerObj_Ser, compN = 2, reduction = "inmf",
plot = T,  group.by = "Species", cols = col_vector)

}
