% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{CellScore_VlnPlot}
\alias{CellScore_VlnPlot}
\title{Cell Score Violin Plot}
\usage{
CellScore_VlnPlot(
  SerObj,
  reduction = "inmf",
  group.by = "orig.ident",
  compN = 1,
  plot = T,
  raster = F,
  pt.size = 0.1,
  cols = NULL
)
}
\arguments{
\item{SerObj}{A Seurat SerObj.}

\item{reduction}{which reduction to use to get cell scores}

\item{plot}{if true returns plot else returns SerObj with CellScore}

\item{raster}{raster or not}

\item{pt.size}{point size passed to featureplot}

\item{cols}{color vector passed to VlnPlot}
}
\value{
plot or SerObj
}
\description{
Plots a violin plot of cell scores from a specific reduction
}
\details{
General principles behind it:
}
\examples{
CellScore_VlnPlot(SerObj = ComboSerObj_Ser, compN = 2, reduction = "inmf",
plot = T, pt.size = .5,  group.by = "Species", cols = col_vector)

}
