% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{cleanmygenes.mouse}
\alias{cleanmygenes.mouse}
\title{clean gene set of commonly unwanted genes}
\usage{
cleanmygenes.mouse(geneV, keepV = NULL)
}
\arguments{
\item{geneV}{mouse gene set}
}
\value{
cleaned mouse gene set
}
\description{
This function cleans the input gene set
}
