% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{closest_1_0s}
\alias{closest_1_0s}
\title{Find the closest multiple of 10 to a given number.}
\usage{
closest_1_0s(x)
}
\arguments{
\item{x}{a numeric input.}
}
\value{
the closest multiple of 10 to \code{x}.
}
\description{
This function takes a numeric input \code{x} and returns the closest multiple of 10 to \code{x}.
}
\examples{
closest_1_0s(12345)
closest_1_0s(6.75)
}
