% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_Seurat.R
\name{CommonLoadingSDAMat}
\alias{CommonLoadingSDAMat}
\title{Combine loading matrices across multiple SDA runs}
\usage{
CommonLoadingSDAMat(sdaRuns)
}
\arguments{
\item{sdaRuns}{A list of SDA runs, each containing a loadings matrix.}
}
\value{
A matrix of loadings that are common across all SDA runs.
}
\description{
This function takes a list of SDA runs and returns a matrix of loadings that are common across all runs.
}
\examples{
## Create a matrix of common loadings across multiple SDA runs
my_loadings <- CommonLoadingSDAMat(sdaRuns = my_sda_runs)

}
