% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{compressSerHD5}
\alias{compressSerHD5}
\title{compressSerHD5}
\usage{
compressSerHD5(
  so = NULL,
  load.path = NULL,
  overwrite = F,
  save.path = NULL,
  updateSerObj = F,
  returnSerObj = F
)
}
\arguments{
\item{so}{A Seurat Object}

\item{load.path}{the path to a seurat .rds file}

\item{overwrite}{default F overwirtes the new hd5 if exists}

\item{save.path}{the path to the desired hd5 seurat file. Default is NULL which then load.path is used to save the new object next to it.}

\item{updateSerObj}{default F if T runs UpdateSeuratObject()}

\item{returnSerObj}{default F if T returns the Seurat object to be used in a pipeline}
}
\value{
Saves HD5 Seurat object to path given
}
\description{
compressSerHD5 will save space on your drives.
}
