% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GausMixMod.R
\name{compute_gmm_threshold}
\alias{compute_gmm_threshold}
\title{Compute the intersection threshold for a two‐component Gaussian mixture}
\usage{
compute_gmm_threshold(params, dataV)
}
\arguments{
\item{params}{A list of mixture parameters, typically `model$parameters` from an \code{Mclust} object,
containing:
\itemize{
  \item \code{mean}: Numeric vector of component means.
  \item \code{variance$sigmasq}: Numeric vector of component variances.
}}

\item{dataV}{A vector of the data used to make gmm}
}
\value{
A single numeric value representing the threshold at which the two Gaussian
  components have equal density (the intersection point).
}
\description{
Given the parameters of a two‐component Gaussian mixture model (means and variances),
this function solves for the point where the two Gaussian density curves intersect.
}
\examples{
\dontrun{
  fit <- Mclust(my_data, G = 2)
  thresh <- compute_gmm_threshold(fit$parameters, dataV=my_data)
}
}
