% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{ComputeARI_Ser}
\alias{ComputeARI_Ser}
\title{Compute Adjusted Rand Index (ARI) for a Seurat Object}
\usage{
ComputeARI_Ser(
  SerObj,
  true.label = "cell_type",
  predicted.label = "seurat_clusters"
)
}
\arguments{
\item{SerObj}{A Seurat object containing the single-cell RNA-seq data.}

\item{true.label}{Character. The metadata label representing the true biological labels (e.g., "cell_type") in `SerObj@meta.data`. Default is `"cell_type"`.}

\item{predicted.label}{Character. The metadata label representing the predicted cluster assignments (e.g., "seurat_clusters") in `SerObj@meta.data`. Default is `"seurat_clusters"`.}
}
\value{
The Adjusted Rand Index (ARI) score. A higher ARI indicates better alignment between the true labels and the predicted clusters.
}
\description{
This function calculates the Adjusted Rand Index (ARI) between the true biological labels (e.g., cell types) and the predicted cluster assignments in a Seurat object. ARI is used to evaluate the similarity between two clustering results.
}
\examples{
# Compute ARI between true cell types and predicted Seurat clusters
ari_score <- ComputeARI_Ser(SerObj, true.label = "cell_type", predicted.label = "seurat_clusters")

# Compute ARI for a different set of predicted clusters
ari_score <- ComputeARI_Ser(SerObj, true.label = "cell_type", predicted.label = "cluster_method_X")

}
