% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{ComputeLISI_Ser}
\alias{ComputeLISI_Ser}
\title{Compute LISI (Local Inverse Simpson's Index) for a Seurat Object}
\usage{
ComputeLISI_Ser(
  SerObj,
  reduction = "pca",
  dims = 1:15,
  batch.label = "batch",
  biological.label = "cell_type"
)
}
\arguments{
\item{SerObj}{A Seurat object containing the single-cell RNA-seq data.}

\item{reduction}{Character. The dimensional reduction method to use for LISI computation (e.g., "pca", "umap"). Default is `"pca"`.}

\item{dims}{Numeric vector. The dimensions of the reduced embedding to use. Default is `1:15`.}

\item{batch.label}{Character. The metadata label representing batch information in `SerObj@meta.data`. Default is `"batch"`.}

\item{biological.label}{Character. The metadata label representing biological information (e.g., cell type) in `SerObj@meta.data`. Default is `"cell_type"`.}
}
\value{
A list containing the average batch LISI score (`batch_LISI`) and the average biological LISI score (`biological_LISI`). A higher batch LISI score indicates better mixing, and a higher biological LISI score indicates better preservation of biological signal.
}
\description{
This function calculates the LISI (Local Inverse Simpson's Index) scores for batch mixing and biological signal preservation based on a specified dimensional reduction (e.g., PCA, UMAP) for a Seurat object.
}
\examples{
# Compute LISI scores for PCA reduction with batch and cell type labels
lisi_scores <- ComputeLISI_Ser(SerObj, reduction = "pca", dims = 1:15, batch.label = "batch", biological.label = "cell_type")

# Compute LISI scores for UMAP reduction
lisi_scores <- ComputeLISI_Ser(SerObj, reduction = "umap", dims = 1:10, batch.label = "batch", biological.label = "cell_type")

}
