% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{convert.Human2Mouse.biomaRt}
\alias{convert.Human2Mouse.biomaRt}
\title{convert.Human2Mouse.biomaRt}
\usage{
convert.Human2Mouse.biomaRt(x)
}
\arguments{
\item{x}{mouse gene set}
}
\value{
converted human gene set
}
\description{
This function converts human genes to mouse genes via biomaRt
}
