% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{convert.Human2Mouse}
\alias{convert.Human2Mouse}
\title{convert.Human2Mouse}
\usage{
convert.Human2Mouse(geneV)
}
\arguments{
\item{geneV}{mouse gene set}
}
\value{
converted human gene set
}
\description{
This function converts human genes to mouse genes via text manipulation e.g. CD8A -> Cd8as
}
