% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{convert.Mouse2Human.biomaRt}
\alias{convert.Mouse2Human.biomaRt}
\title{convert.Mouse2Human.biomaRt}
\usage{
convert.Mouse2Human.biomaRt(x)
}
\arguments{
\item{x}{mouse gene set}
}
\value{
converted human gene set
}
\description{
This function converts mouse genes to human genes via biomaRt
}
