% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoMX.R
\name{convert_GeoMxSet_to_Seurat}
\alias{convert_GeoMxSet_to_Seurat}
\title{Convert GeoMxSet Object to Seurat}
\usage{
convert_GeoMxSet_to_Seurat(
  x,
  normData,
  ident = NULL,
  coordinates = NULL,
  sequencingMetrics = NULL,
  QCMetrics = "QCFlags"
)
}
\arguments{
\item{x}{A GeoMxSet object.}

\item{normData}{Name of the normalized expression matrix (e.g., "exprs").}

\item{ident}{Optional identity column name in sample metadata to set Seurat identities.}

\item{coordinates}{Optional character vector of length 2 indicating x and y coordinate column names.}

\item{sequencingMetrics}{Optional vector of sequencing metric column names to retain in @misc.}

\item{QCMetrics}{Optional QC metrics column name to retain in @misc.}
}
\value{
A Seurat object.
}
\description{
Convert GeoMxSet Object to Seurat
}
