% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{count_overlap}
\alias{count_overlap}
\title{Counts the overlap of items in rows per column feature}
\usage{
count_overlap(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
rturns a count matrix
}
\description{
This function takes in a list of ggplots and makes a GIF
}
