% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{create_gif_LS}
\alias{create_gif_LS}
\title{Make a GIF from plot list}
\usage{
create_gif_LS(plot_list, filename = "animation.gif", delay = 100, res = 72)
}
\arguments{
\item{plot_list}{list of ggplots}

\item{filename}{path and filename of gif}
}
\value{
saves filename gif
}
\description{
This function takes in a list of ggplots and makes a GIF
}
