% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{create_grouped_plot}
\alias{create_grouped_plot}
\title{Create a grouped plot with percentage labels in the legend.}
\usage{
create_grouped_plot(ComboSerObj, group_feature, color_values)
}
\arguments{
\item{ComboSerObj}{A Seurat object containing the data.}

\item{group_feature}{The feature based on which the data will be grouped and plotted.}

\item{color_values}{A vector of color values for the groups.}
}
\value{
A ggplot object with grouped points and legend labels representing percentages.
}
\description{
This function generates a grouped plot for the specified feature in a Seurat object, 
adding percentage labels to the legend to represent the proportion of total points 
in each category.
}
\examples{
\dontrun{
# Example usage:
# gg <- create_grouped_plot(ComboSerObj, group_feature = "IsAlphaBeta", 
#                           color_values = c("dodgerblue", "maroon"))
}
}
