% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{create_qq_plot}
\alias{create_qq_plot}
\title{Create a QQ Plot}
\usage{
create_qq_plot(
  x,
  y,
  x_title = "Theoretical Quantiles",
  y_title = "Sample Quantiles",
  title = "QQ Plot",
  theme = theme_bw(base_size = 14)
)
}
\arguments{
\item{x}{Not used (for flexibility).}

\item{y}{The sample data for which you want to create the QQ plot.}

\item{x_title}{The title for the x-axis.}

\item{y_title}{The title for the y-axis.}

\item{title}{The title for the plot.}
}
\value{
A ggplot object representing the QQ plot.
}
\description{
This function generates a quantile-quantile (QQ) plot using ggplot2.
}
\examples{
set.seed(123)
sample_data <- rnorm(100)
qq_plot <- create_qq_plot(NULL, sample_data, "Custom X-axis Title", "Custom Y-axis Title", "Custom QQ Plot Title")
print(qq_plot)

}
