% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{create_upset_plot}
\alias{create_upset_plot}
\title{Create an UpSet plot from a list of sets}
\usage{
create_upset_plot(
  set_list,
  output_filename = NULL,
  image_format = "png",
  do.sqr = T,
  plot.nrow = 2,
  plot.ncol = 2,
  plot.heights = c(1, 2),
  plot.widths = c(1, 0.8),
  save.height = 3.5,
  save.width = 3
)
}
\arguments{
\item{set_list}{A list of sets to visualize.}

\item{output_filename}{The filename for the output image.}

\item{image_format}{The format in which to save the image (default: "svg").}

\item{do.sqr}{Should the y-axis of the main plot and overlap plot be square-root transformed? (default: TRUE)}
}
\value{
A ggplot2 UpSet plot.
}
\description{
This function generates an UpSet plot from a list of sets and saves it as an image.
}
\examples{
# Example usage:
create_upset_plot(list("A" = c("1", "2", "3"), "B" = c("2", "3", "4")), "upset_plot.svg")


}
