% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{CreateProteinHeatmap}
\alias{CreateProteinHeatmap}
\title{Create a protein heatmap}
\usage{
CreateProteinHeatmap(
  SerObj,
  proteinList,
  labelColumn,
  prefix,
  size,
  fontsize,
  titlefontsize,
  fullorder
)
}
\arguments{
\item{SerObj}{A Seurat object.}

\item{proteinList}{A list of proteins to plot.}

\item{labelColumn}{Column name to use for labeling rows.}

\item{prefix}{Prefix for plot title.}

\item{size}{Size of plot.}

\item{fontsize}{Font size for labels.}

\item{titlefontsize}{Font size for plot title.}

\item{fullorder}{}
}
\value{
A heatmap of protein expression.
}
\description{
Create a protein heatmap
}
