% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_tables.R
\name{discretize_table}
\alias{discretize_table}
\title{discretize_table}
\usage{
discretize_table(tbl = NULL)
}
\arguments{
\item{tbl, }{numbers.}
}
\value{
histo_numers
}
\description{
discretize a table; if tbl>0 then 1 else 0
}
