% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlotFxs.R
\name{DotPlot2}
\alias{DotPlot2}
\title{Enhanced Dot Plot Visualization for Seurat Objects}
\usage{
DotPlot2(
  SerObj,
  group.by,
  layer = "data",
  assay = "RBA",
  markerVec,
  rowsplit = NULL,
  columnsplit = NULL,
  size = 6,
  coldend = FALSE,
  rowdend = FALSE,
  coldendside = "bottom",
  rowdendside = "left",
  fontsize = 12,
  titlefontsize = 14,
  gap = 0
)
}
\arguments{
\item{SerObj}{A Seurat object containing the data to be visualized.}

\item{group.by}{The name of the metadata column in `SerObj` to group the data by.}

\item{layer}{The layer from which to retrieve data. Common layers include "data", "scale.data", and "counts".}

\item{assay}{The name of the assay to use for data retrieval. Default is "RBA".}

\item{markerVec}{A vector of marker genes to include in the plot.}

\item{rowsplit}{Optional. A vector specifying how to split rows in the plot. NULL means no splitting.}

\item{columnsplit}{Optional. A vector specifying how to split columns in the plot. NULL means no splitting.}

\item{size}{The size of the dots in the plot.}

\item{coldend}{Logical, whether to include a column dendrogram.}

\item{rowdend}{Logical, whether to include a row dendrogram.}

\item{coldendside}{The side on which to draw the column dendrogram ("top" or "bottom").}

\item{rowdendside}{The side on which to draw the row dendrogram ("left" or "right").}

\item{fontsize}{The font size for text in the plot.}

\item{titlefontsize}{The font size for the plot title.}

\item{gap}{The gap between dots in the plot.}
}
\value{
Generates a dot plot visualization based on the provided Seurat object and parameters.
}
\description{
This function provides an enhanced dot plot visualization for Seurat objects,
allowing for flexible grouping, assay selection, and dendrogram inclusion options.
It averages expression data by specified groups and then utilizes a customized
plotting function to display the results.
}
\examples{
DotPlot2(mySeuratObject, group.by = "ident", layer = "data", assay = "RNA",
         markerVec = c("GeneA", "GeneB", "GeneC"))

}
