% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabKey_based.R
\name{download_CITEseq}
\alias{download_CITEseq}
\title{download_CITEseq}
\usage{
download_CITEseq(barcodePrefix, outPath = "./data/CITEseq")
}
\arguments{
\item{barcodePrefix}{barcodePrefix e.g. 191341}

\item{outPath}{A path to save the CITEseq files}
}
\value{
saves calls in outPath
}
\description{
Download the hashing calls for a seurat object based on LoupeId
}
