% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabKey_based.R
\name{download_SeuratHashingCalls}
\alias{download_SeuratHashingCalls}
\title{download_SeuratHashingCalls}
\usage{
download_SeuratHashingCalls(
  loupeId,
  outPath = "./",
  lkBaseUrl = "https://prime-seq.ohsu.edu",
  lkDefaultFolder = "/Labs/Bimber"
)
}
\arguments{
\item{loupeId}{LoupID e.g. 169823}

\item{outPath}{A path to save the hash files}

\item{lkBaseUrl}{setting of labkey}

\item{lkDefaultFolder}{setting of labkey}
}
\value{
saves hashing calls in outPath as .3.hashing.calls.txt
}
\description{
Download the hashing calls for a seurat object based on LoupeId
}
