% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{DownSample_SerObj_perLevel}
\alias{DownSample_SerObj_perLevel}
\title{DownSample_SerObj_perLevel}
\usage{
DownSample_SerObj_perLevel(SerObj, featureSplit = NULL, totalCell = 300)
}
\arguments{
\item{SerObj}{A Seurat Object}

\item{featureSplit}{a feature of seurat object to split. "BarcodePrefix" # this is a good way to sample per seq run if possible}

\item{totalCell}{0 # is the default way to downsample by min common cells}
}
\value{
A subseted downsampled Seurat object.
}
\description{
downsample per feature levels
}
