% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkRNAseq.R
\name{draw_volcano_DEseq}
\alias{draw_volcano_DEseq}
\title{Draw a volcano plot for one contrast derived from DEseq}
\usage{
draw_volcano_DEseq(contrast_df, Title = NULL)
}
\arguments{
\item{contrast_df}{Data frame with columns: log2FoldChange, padj, gene, contrast_name}

\item{Title}{Title text or NULL}
}
\value{
ggplot object
}
\description{
Draw a volcano plot for one contrast derived from DEseq
}
