% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalizations.R
\name{dropsimNorm}
\alias{dropsimNorm}
\title{dropsimNorm}
\usage{
dropsimNorm(
  GeneExprMat,
  featuresToUse,
  minLibrarySize = 0,
  maxLibrarySize = 10,
  center = FALSE,
  scale = TRUE,
  gene_subset = 1
)
}
\arguments{
\item{GeneExprMat}{A matrix rows of genes and cols of cells}
}
\value{
A quantile-based normalized matrix
}
\description{
passes to dropsim::normaliseDGE() fx
}
