% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalizations.R
\name{dropsimNormV2}
\alias{dropsimNormV2}
\title{dropsimNormV2 speeds up using sparse matrix fx as much as possible}
\usage{
dropsimNormV2(
  GeneExprMat,
  featuresToUse,
  minLibrarySize = 0,
  maxLibrarySize = 10,
  center = FALSE,
  scale = TRUE,
  gene_subset = 1
)
}
\arguments{
\item{GeneExprMat}{A matrix rows of genes and cols of cells}
}
\value{
A quantile-based normalized matrix
}
\description{
calls normaliseDGEv2 instead of dropsim::normaliseDGE()
}
