% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{DualFeatGate}
\alias{DualFeatGate}
\title{Perform dual-feature gating on a Seurat object}
\usage{
DualFeatGate(
  SerObject,
  feat1 = NULL,
  feat2 = NULL,
  thr1 = 0,
  thr2 = 0,
  dir1 = "pos",
  dir2 = "pos",
  cols = c("maroon", "gray", "blue", "forestgreen"),
  plotDimplot = T,
  returnSerObj = T,
  reduction = "umap",
  label = F,
  label.size = 6,
  repel = T,
  raster = F
)
}
\arguments{
\item{SerObject}{A Seurat object to perform gating on.}

\item{feat1}{A character string representing the name of the first feature to gate on. Defaults to NULL.}

\item{feat2}{A character string representing the name of the second feature to gate on. Defaults to NULL.}

\item{thr1}{A numeric threshold value for the first feature. Defaults to 0.}

\item{thr2}{A numeric threshold value for the second feature. Defaults to 0.}

\item{dir1}{A character string representing the direction of the threshold for the first feature, either "pos" or "neg". Defaults to "pos".}

\item{dir2}{A character string representing the direction of the threshold for the second feature, either "pos" or "neg". Defaults to "pos".}

\item{plotDimplot}{A logical value indicating whether to plot the result using Seurat's DimPlot function. Defaults to TRUE.}

\item{returnSerObj}{A logical value indicating whether to return the modified Seurat object or just the gated cell names. Defaults to TRUE.}
}
\value{
A Seurat object with the compGate metadata column added if \code{returnSerObj} is TRUE, otherwise a character vector of gated cell names.
}
\description{
Perform dual-feature gating on a Seurat object
}
