% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{FeaturePlot_cust}
\alias{FeaturePlot_cust}
\title{Generate a FeaturePlot_cust which is a scatter plot with gene expression as color scale}
\usage{
FeaturePlot_cust(
  SerObject,
  dim1,
  dim2,
  pt.size = 1,
  group.by,
  Feat,
  colors = c("white", "gray", "gold", "red", "maroon"),
  base_size = 14
)
}
\arguments{
\item{SerObject}{A Seurat object containing the data to plot}

\item{dim1}{Name of the first feature to plot on the x-axis}

\item{dim2}{Name of the second feature to plot on the y-axis}

\item{pt.size}{Point size for the plot (default: 1)}

\item{group.by}{Name of the grouping variable for the plot (optional)}

\item{Feat}{Name of the feature to use for coloring the plot}

\item{colors}{Vector of colors for the color gradient (default: c('white', 'gray', 'gold', 'red', 'maroon'))}

\item{base_size}{Base font size for the plot (default: 14)}
}
\value{
A ggplot object
}
\description{
This function generates a scatter plot of the specified dimensions (features) of a Seurat object,
colored by expression values of the specified feature.
}
