% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Spatial_based.R
\name{FilterSpatial}
\alias{FilterSpatial}
\title{Filter Spatial data and plot before and after cell loss}
\usage{
FilterSpatial(
  SerObj,
  nCount_Spatial_max = 1000,
  nFeature_Spatial_min = 20,
  nFeature_Spatial_max = 700,
  percent.mt_max = 10
)
}
\arguments{
\item{SerObj}{A SerObj}

\item{nCount_Spatial_max}{pnCount_Spatial_max}

\item{nFeature_Spatial_min}{pnCount_Spatial_max}

\item{nFeature_Spatial_max}{pnCount_Spatial_max}

\item{percent.mt_max}{pnCount_Spatial_max}
}
\value{
a Seurat object for Spatial SlideSeq data
}
\description{
Filter Spatial data and plot before and after cell loss
}
