% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Find_FOXP3}
\alias{Find_FOXP3}
\title{Find_FOXP3
Identify FOXP3+ cells}
\usage{
Find_FOXP3(SerObj, plot = FALSE)
}
\arguments{
\item{SerObj}{A Seurat object containing the data.}
}
\value{
A Seurat object with the IsFOXP3 metadata column added
}
\description{
Find_FOXP3
Identify FOXP3+ cells
}
