% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{find_pseudotime}
\alias{find_pseudotime}
\title{Find Pseudotime}
\usage{
find_pseudotime(sample.point, pseudotime)
}
\arguments{
\item{sample.point}{A numeric vector representing the sample point.}

\item{pseudotime}{A matrix or data frame containing pseudotime values.}
}
\value{
The index of the pseudotime corresponding to the closest match to the given sample point.
}
\description{
This function finds the pseudotime corresponding to a given sample point.
}
\examples{
sample.point <- c(0.5, 0.3, 0.7)
pseudotime <- matrix(c(0.1, 0.4, 0.8, 0.2, 0.3, 0.6), ncol = 3)
find_pseudotime(sample.point, pseudotime)

}
