% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{Get.Gene_Info}
\alias{Get.Gene_Info}
\title{Get.Gene_Info}
\usage{
Get.Gene_Info(
  gene_names,
  dataset = "hsapiens_gene_ensembl",
  attributes = c("external_gene_name", "chromosome_name", "start_position",
    "end_position", "description")
)
}
\arguments{
\item{gene_names}{gene set}

\item{dataset}{default hsapiens_gene_ensembl, try mmusculus_gene_ensembl}

\item{attributes}{attributes to grab}
}
\value{
converted human gene set
}
\description{
This function get gene info via biomaRt
}
