% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MitoGenes.R
\name{get_MitoGeneSet}
\alias{get_MitoGeneSet}
\title{get_MitoGeneSet}
\usage{
get_MitoGeneSet(Name = NULL)
}
\arguments{
\item{Name}{of of either c("mito1"); mito1 = www.genedx.com list.}
}
\value{
a character vector of genes
}
\description{
returns a set of genes.
}
\details{
https://www.genedx.com/wp-content/uploads/crm_docs/Mito-Gene-List.pdf
}
