% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{getFeatWindows}
\alias{getFeatWindows}
\title{Get feature windows}
\usage{
getFeatWindows(SerObject, feat, group.by)
}
\arguments{
\item{SerObject}{A Seurat object.}

\item{feat}{A character vector representing the feature name.}

\item{group.by}{A character vector representing the group to use for grouping the data.}
}
\value{
A list with two elements:
\describe{
  \item{winDF}{A table with the counts of cells falling into each window, for each group.}
  \item{vlines}{A numeric vector with the vertical lines representing the boundaries of each window.}
}
}
\description{
This function generates feature windows for a given feature and groups in a Seurat object.
}
