% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{getGeneHomology}
\alias{getGeneHomology}
\title{Retrieve Homology Information from Ensembl}
\usage{
getGeneHomology(
  mirror = "https://uswest.ensembl.org",
  species_list = c("mmulatta")
)
}
\arguments{
\item{mirror}{A character string specifying the Ensembl mirror to use. Default is `"https://uswest.ensembl.org"`.}

\item{species_list}{A character vector of species names (in Ensembl format) for which homology data should be retrieved. 
Default includes `"mmusculus"`, `"mmulatta"`, `"ppaniscus"`, `"ptroglodytes"`, `"ggorilla"`, `"nleucogenys"`, `"hsapiens"`, and `"cjacchus"`.}
}
\value{
A named list where each element contains a data frame with homology information for a given species.
}
\description{
This function queries Ensembl using the `biomaRt` package to retrieve ortholog information 
for multiple species relative to human genes. It returns a list where each element contains 
the homology data for a given species.
}
\examples{
\dontrun{
homology_data <- getGeneHomology()
head(homology_data$mmusculus)
}
}
