% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_Analysis.R
\name{GetTopKEGG}
\alias{GetTopKEGG}
\title{GetTopKEGG}
\usage{
GetTopKEGG(df = NULL, top.by = "Selected_PV", type = "double", ntop = 20)
}
\arguments{
\item{df}{Data frame containing KEGG pathways.}

\item{top.by}{Label used to select the top.}

\item{type}{Type of selected p-values. Available options are: "double", negatives and positive, "positive", and "negative".}

\item{ntop}{Number of top values.}
}
\value{
Data frame containing the top values.
}
\description{
GetTopKEGG
}
