% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_Analysis.R
\name{GetTopPathways}
\alias{GetTopPathways}
\title{GetTopPathways}
\usage{
GetTopPathways(dfs = NULL, ntop = 20)
}
\arguments{
\item{dfs}{List of dataframes containing KEGG pathways}

\item{ntop}{Number of top pathways to select.}
}
\value{
List containing the selected pathways and their ranks (e.g. number of dfs containing the pathway).
}
\description{
GetTopPathways
}
