% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{gg_barplot_2waytable}
\alias{gg_barplot_2waytable}
\title{gg_barplot_table, aka OOSAP::PlotMyTable}
\usage{
gg_barplot_2waytable(
  MyTable,
  Title = "",
  legend.position = "bottom",
  PlotCombo = F
)
}
\arguments{
\item{MyTable}{A 2way table (base R table())}

\item{Title}{A title else ''}

\item{legend.position}{legend position defautl "bottom"}

\item{PlotCombo}{if T plots % and count plots together}
}
\value{
A ggplot
}
\description{
Takes a table and plots a gg barplot
}
