% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_tables.R
\name{gg_barplot_table2}
\alias{gg_barplot_table2}
\title{gg_barplot_table2}
\usage{
gg_barplot_table2(TBL, Title = NULL, gg_theme = NULL, col_vector = NULL)
}
\arguments{
\item{TBL}{A table (base R table())}

\item{Title}{A title or NULL}

\item{gg_theme}{a ggplot theme or NULL}

\item{col_vector}{a vector of colors or NULL}
}
\value{
A ggplot
}
\description{
Takes a table and plots a gg barplot
}
