% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_tables.R
\name{gg_pie_table}
\alias{gg_pie_table}
\title{gg_pie_table}
\usage{
gg_pie_table(TBL = NULL, cols = col_vector, legend.position = "none")
}
\arguments{
\item{TBL}{A table or data frame containing the data for the pie chart.}

\item{cols}{A vector of colors to be used for filling the pie slices.}

\item{legend.position}{The position of the legend in the plot.}
}
\description{
This function creates a pie chart with labels and frequency values using ggplot2.
}
