% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{heatmap_table}
\alias{heatmap_table}
\title{heatmap_table}
\usage{
heatmap_table(
  TBL,
  clust.meth = "ward.D2",
  clust.dist = "euclidean",
  Title = "",
  plotTable = F
)
}
\arguments{
\item{TBL}{A table (base R table())}

\item{clust.meth}{default is "ward.D2", see pheatmap for more options}

\item{clust.dist}{default is "euclidean", see pheatmap for more options}

\item{Title}{A title else ''}

\item{plotTable}{default is F, if T, it will plot an actual table}
}
\value{
A pheatmap heatmap or a plotted table
}
\description{
Takes a table and plots it. Either as a table or a pheatmap heatmap with clustering
}
