% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_Seurat.R
\name{ImputeSDA2SerV3}
\alias{ImputeSDA2SerV3}
\title{This function to add imputed scores from gene loading, V3 is one component at a time selecting topN genes}
\usage{
ImputeSDA2SerV3(
  SerObj,
  sda_loadings,
  keepComps = NULL,
  sdaObjID = "",
  plot = F,
  doAsinh = T,
  TopNpos = 20,
  TopNneg = 20,
  MetaExtraName = ""
)
}
\arguments{
\item{SerObj}{A seurat Object, with assay RNA and slot data}

\item{sda_loadings}{matrix of sda loadings}

\item{keepComps}{if NULL all else numeric or char vec}

\item{sdaObjID}{id name to apped to comp names}

\item{plot}{to plot or not}
}
\value{
seurat obj with new SDA score comps in metadata
}
\description{
This function to add imputed scores from gene loading, V3 is one component at a time selecting topN genes
}
