% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{intersect_perc}
\alias{intersect_perc}
\title{Compute the percentage of intersection between two character vectors}
\usage{
intersect_perc(vector1, vector2)
}
\arguments{
\item{vector1}{A character vector.}

\item{vector2}{A character vector.}
}
\value{
The percentage of intersection between the two character vectors.
}
\description{
This function calculates the percentage of intersection between two character vectors.
It converts the vectors to sets, finds the intersection, and computes the percentage
based on the size of the intersection relative to the first vector.
}
\examples{
vector1 <- c("A", "B", "C", "D", "E")
vector2 <- c("C", "D", "E", "F", "G")
intersect_perc(vector1, vector2)
}
