% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{is_in_range}
\alias{is_in_range}
\title{Check if a value falls within a distribution range defined by a set of numbers}
\usage{
is_in_range(value, my_set)
}
\arguments{
\item{value}{The value to be checked}

\item{my_set}{A set of numbers defining the distribution range}
}
\value{
A logical value, TRUE if the value falls within the distribution range, FALSE otherwise
}
\description{
This function checks if a value falls within the distribution range defined by a set
of numbers, represented by the minimum and maximum values of the set.
}
\examples{
my_set <- c(10, 20, 30, 40, 50)
is_in_range(25, my_set)

}
