% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Spatial_based.R
\name{LoadSCE_ReturnSer}
\alias{LoadSCE_ReturnSer}
\title{load a h5ad SCE object and convert to seurat}
\usage{
LoadSCE_ReturnSer(
  SCE_h5ad_file,
  ReturnRawSer = T,
  SER_rds_file,
  verbose = T,
  nCount_Spatial_max = 1000,
  nFeature_Spatial_min = 20,
  nFeature_Spatial_max = 700,
  percent.mt_max = 10,
  LogNorm = F
)
}
\arguments{
\item{SCE_h5ad_file}{path to h5ad file}

\item{ReturnRawSer}{if T no normalization or filtering is done}

\item{SER_rds_file}{path to save rds of ser obj}

\item{verbose}{to print and plot or not T/F}
}
\value{
a Seurat object for Spatial SlideSeq data
}
\description{
load a h5ad SCE object and convert to seurat
}
