% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{make_RNA_heatmap}
\alias{make_RNA_heatmap}
\title{Make an RNA heatmap}
\usage{
make_RNA_heatmap(
  seuratObj,
  markerVec,
  labelColumn,
  rowsplit = NULL,
  columnsplit = NULL,
  size = 6,
  clus_cols = TRUE,
  show_column_dend = T,
  clus_rows = TRUE,
  show_row_dend = T,
  fontsize = 10,
  titlefontsize = 10,
  pairedList2 = NULL,
  asGG = F,
  column_names_side = "bottom",
  column_dend_side = "bottom",
  row_names_side = "left",
  row_dend_side = "left",
  assays = "RNA",
  useCDnames = T
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{markerVec}{A vector of marker genes.}

\item{labelColumn}{Column name to use for labeling rows.}

\item{rowsplit}{Column name to split rows by.}

\item{columnsplit}{Column name to split columns by.}

\item{size}{Size of plot.}

\item{fontsize}{Font size for labels.}

\item{titlefontsize}{Font size for plot title.}

\item{asGG}{returns GGplot not list if true}

\item{coldendside}{Dendrogram on the column side.}

\item{rowdendside}{Dendrogram on the row side.}
}
\value{
A heatmap of RNA expression.
}
\description{
Make an RNA heatmap
}
