% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{make_RNA_heatmap2}
\alias{make_RNA_heatmap2}
\title{Plot RNA heatmap v2, no re-normalization of the data}
\usage{
make_RNA_heatmap2(
  seuratObj,
  labelColumn = "ClusterNames_0.2",
  markerVec = NULL,
  assay = "RNA",
  clus_cols = TRUE,
  show_column_dend = T,
  clus_rows = TRUE,
  show_row_dend = T,
  rowsplit = NULL,
  columnsplit = NULL,
  show_heatmap_legend = T,
  size = 4,
  column_names_rot = 45,
  fontsize = 8,
  titlefontsize = 12,
  coldendside = "bottom",
  asGG = T
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{labelColumn}{Column for cell labels.}

\item{markerVec}{Marker genes to use.}

\item{assay}{Assay type.}

\item{clus_cols}{Cluster columns.}

\item{show_column_dend}{Show column dendrogram.}

\item{clus_rows}{Cluster rows.}

\item{show_row_dend}{Show row dendrogram.}

\item{rowsplit}{Split rows.}

\item{columnsplit}{Split columns.}

\item{show_heatmap_legend}{Show heatmap legend.}

\item{size}{Plot size.}

\item{column_names_rot}{Rotation for column names.}

\item{fontsize}{Font size for labels.}

\item{titlefontsize}{Font size for plot title.}

\item{coldendside}{Position of column dendrogram.}

\item{asGG}{Output as ggplot object.}
}
\value{
A heatmap of RNA expression.
}
\description{
Plot RNA heatmap v2, no re-normalization of the data
}
