% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDimScaling.R
\name{MDSmyDF}
\alias{MDSmyDF}
\title{MDS Plot of Data Frame}
\usage{
MDSmyDF(dfx, labelsDF, factorV, title = "MDS Plot", col_vector, returnMDS = F)
}
\arguments{
\item{dfx}{A data frame with samples as rows and variables as columns.}

\item{labelsDF}{A data frame with the labels for each sample in the same order as dfx.}

\item{factorV}{A vector containing the factor variable to use for coloring the points.}

\item{title}{A character string for the title of the plot.}

\item{col_vector}{A vector of colors to use for coloring the points.}

\item{returnMDS}{A logical value indicating whether to return the MDS coordinates as a data frame.
Default is FALSE.}
}
\value{
If returnMDS is TRUE, returns a data frame of the MDS coordinates. Otherwise, returns
  a ggplot object.
}
\description{
Performs multidimensional scaling (MDS) on a data frame and generates a plot of the
results. The function uses the ggplot2 package to create a scatterplot of the MDS
coordinates with colored points based on a factor variable. The ggrepel package is used
to add labels to the points without overlap.
}
\examples{
data(iris)
MDSmyDF(dfx = iris[, 1:4], labelsDF = iris[, 5], factorV = iris[, 5], title = "Iris Data MDS Plot", 
  col_vector = c("#FF0000", "#00FF00", "#0000FF"))
}
