% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Most_Common_Clones}
\alias{Most_Common_Clones}
\title{Most_Common_Clones
Function to compute the most common clones for a TCR feature in a Seurat object}
\usage{
Most_Common_Clones(SerObj, tcr_feature, top_n = 10)
}
\arguments{
\item{SerObj}{A Seurat object containing the data.}

\item{tcr_feature}{A metadata feature e.g. TRA_J}

\item{top_n}{a numerical value, default 10}
}
\description{
Most_Common_Clones
Function to compute the most common clones for a TCR feature in a Seurat object
}
