% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Most_Common_Clones_2_Matrix}
\alias{Most_Common_Clones_2_Matrix}
\title{Most_Common_Clones_2_Matrix
A function that takes in a list, from Most_Common_Clones_Per()}
\usage{
Most_Common_Clones_2_Matrix(top_clones_list)
}
\arguments{
\item{top_clones_list}{A list of data frames with Clone and Count.Freq}
}
\description{
Most_Common_Clones_2_Matrix
A function that takes in a list, from Most_Common_Clones_Per()
}
