% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Most_Common_Clones_Per}
\alias{Most_Common_Clones_Per}
\title{Most_Common_Clones_Per
Function to compute the most common clones for a TCR feature in a Seurat object, grouped by a second feature}
\usage{
Most_Common_Clones_Per(SerObj, tcr_feature, grouping_feature, top_n = 10)
}
\arguments{
\item{SerObj}{A Seurat object containing the data.}

\item{tcr_feature}{A metadata feature e.g. TRA_J}

\item{grouping_feature}{A metadata feature to group by e.g. Tissue}

\item{top_n}{a numerical value, default 10}
}
\description{
Most_Common_Clones_Per
Function to compute the most common clones for a TCR feature in a Seurat object, grouped by a second feature
}
