% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMembrane_based.R
\name{NormalizeAndScaleV2}
\alias{NormalizeAndScaleV2}
\title{NormalizeAndScaleV2 a customized version from CellMembrane}
\usage{
NormalizeAndScaleV2(
  seuratObj,
  variableFeatureSelectionMethod = "vst",
  nVariableFeatures = 2000,
  mean.cutoff = c(0.0125, 3),
  dispersion.cutoff = c(0.5, Inf),
  block.size = 1000,
  normalization.method = "LogNormalize",
  Regress_basic = F,
  Regress_pmito = T
)
}
\arguments{
\item{seuratObjects, }{A named list of Seurat object.}
}
\value{
Normalized and scaled seurat object
}
\description{
This is the entry point for processing raw scRNAseq data with Seurat & CellMembrane.
}
