% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalizations.R
\name{normalizeMultipleMatrices}
\alias{normalizeMultipleMatrices}
\title{Normalize multiple matrices using quantile normalization}
\usage{
normalizeMultipleMatrices(matrices)
}
\arguments{
\item{matrices}{A list of matrices}
}
\value{
A list of normalized sparse matrices
}
\description{
Normalize multiple matrices using quantile normalization
}
