% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCAmyDF}
\alias{PCAmyDF}
\title{Principal Component Analysis (PCA) Plot}
\usage{
PCAmyDF(
  dfx,
  labels,
  factorV,
  title = "PCA Plot",
  scale,
  center,
  col_vector,
  namePointBL = F
)
}
\arguments{
\item{dfx}{a numeric matrix containing the data to be plotted}

\item{labels}{a character vector containing the labels for each sample in \code{dfx}}

\item{factorV}{a factor or character vector indicating the group each sample belongs to}

\item{title}{a character string specifying the title of the plot}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance}

\item{center}{a logical value indicating whether the variables should be centered}

\item{col_vector}{a vector of color names to be used for plotting the samples}

\item{namePointBL}{a logical value indicating whether to include labels for each point}
}
\value{
If \code{returnPCA} is TRUE, the function returns a list containing the PCA results. Otherwise, the function generates a PCA plot.
}
\description{
This function generates a PCA plot using the prcomp function in R.
}
\examples{
data("iris")
PCAmyDF(iris[,1:4], iris$Species, iris$Species, "Iris dataset", TRUE, TRUE, c("red", "blue", "green"), TRUE)

}
