% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Radar.R
\name{Perc2Radar}
\alias{Perc2Radar}
\title{Perc2Radar}
\usage{
Perc2Radar(
  PercDF = NULL,
  Features = NULL,
  col_vector = col_vector,
  plotLegend = T
)
}
\arguments{
\item{PercDF}{A data frame containing percentage data.}

\item{Features}{A character vector specifying the features (columns) to be included in the radar chart.}

\item{col_vector}{A vector of colors to be used for plotting the radar chart.}

\item{plotLegend}{A logical value indicating whether to plot the legend.}
}
\description{
This function creates a radar chart based on the percentage data provided in a data frame.
}
