% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{PercentExpressed}
\alias{PercentExpressed}
\title{Calculate the percent expressed cells for each group in a Seurat object}
\usage{
PercentExpressed(so, group.by, features, plot_perHM = F)
}
\arguments{
\item{so}{A Seurat object}

\item{group.by}{A meta feature used to group the cells}

\item{features}{A vector of features (genes) to calculate the percent expressed cells for}

\item{plot_perHM}{A logical indicating whether to plot a heatmap of the results}
}
\value{
A data frame with the percent expressed cells for each group and feature
}
\description{
Calculate the percent expressed cells for each group in a Seurat object
}
