% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_Seurat.R
\name{PercExprThrTab}
\alias{PercExprThrTab}
\title{PercExprThrTab}
\usage{
PercExprThrTab(
  SeuratObj,
  GeneNames = NULL,
  CutThresh = NULL,
  MetaDataName = NULL,
  slot = "data",
  Quant = 0.1,
  assay = "RNA"
)
}
\arguments{
\item{SeuratObj}{A Seurat object.}

\item{GeneNames}{A character vector specifying the gene names to calculate the percentage for.}

\item{CutThresh}{A numeric value specifying the threshold for gene expression. If NULL, quantile thresholding will be used.}

\item{MetaDataName}{A character string specifying the metadata column name to group the cells.}

\item{slot}{The slot name to retrieve the data from. Default is "data".}

\item{Quant}{A numeric value specifying the quantile threshold to use if CutThresh is NULL.}

\item{assay}{The assay name to retrieve the data from. If NULL, the default assay of the Seurat object is used.}
}
\description{
This function calculates the percentage of cells expressing each gene above a given threshold in each metadata group of a Seurat object.
}
