% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_PieDonut.R
\name{PieDonut2}
\alias{PieDonut2}
\title{Plot a pie and donut chart}
\usage{
PieDonut2(
  data,
  mapping,
  start = getOption("PieDonut.start", 0),
  addPieLabel = TRUE,
  addDonutLabel = TRUE,
  showRatioDonut = TRUE,
  showRatioPie = TRUE,
  ratioByGroup = TRUE,
  showRatioThreshold = getOption("PieDonut.showRatioThreshold", 0.02),
  labelposition = getOption("PieDonut.labelposition", 2),
  labelpositionThreshold = 0.1,
  r0 = getOption("PieDonut.r0", 1.2),
  r1 = getOption("PieDonut.r1", 1),
  r2 = getOption("PieDonut.r2", 1.2),
  explode = NULL,
  selected = NULL,
  explodePos = 0.1,
  color = "white",
  pieAlpha = 0.8,
  donutAlpha = 1,
  maxx = NULL,
  showPieName = TRUE,
  showDonutName = FALSE,
  title = NULL,
  pieLabelSize = 4,
  donutLabelSize = 3,
  titlesize = 5,
  explodePie = TRUE,
  explodeDonut = FALSE,
  use.label = TRUE,
  use.labels = TRUE,
  family = getOption("PieDonut.family", "")
)
}
\arguments{
\item{data}{Data to plot.}

\item{mapping}{Mapping for colors.}

\item{start}{Starting angle for pie chart.}

\item{addPieLabel}{Add labels to pie chart.}

\item{addDonutLabel}{Add labels to donut chart.}

\item{showRatioDonut}{Show ratios in donut chart.}

\item{showRatioPie}{Show ratios in pie chart.}

\item{ratioByGroup}{Show ratio by group.}

\item{showRatioThreshold}{Threshold for showing ratios.}

\item{labelposition}{Label position.}

\item{labelpositionThreshold}{Threshold for label position.}

\item{r0}{Inner radius of donut chart.}

\item{explode}{Explode pie chart.}

\item{selected}{Selected data to highlight.}

\item{explodePos}{Explode position.}

\item{color}{Color for pie chart.}

\item{pieAlpha}{Alpha for pie chart.}

\item{donutAlpha}{Alpha for donut chart.}

\item{maxx}{Maximum value for x axis.}

\item{showPieName}{Show name for pie chart.}

\item{showDonutName}{Show name for donut chart.}

\item{title}{Plot title.}

\item{pieLabelSize}{Font size for pie chart labels.}

\item{donutLabelSize}{Font size for donut chart labels.}

\item{titlesize}{Font size for plot title.}

\item{explodePie}{Explode pie chart.}

\item{explodeDonut}{Explode donut chart.}

\item{use.label}{Use label.}

\item{use.labels}{Use labels.}

\item{family}{Font family.}
}
\value{
A pie and donut chart.
}
\description{
Plot a pie and donut chart
}
