% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{plot2Dredux}
\alias{plot2Dredux}
\title{Plot 2D Dimensionality Reduction}
\usage{
plot2Dredux(
  plotDF,
  centers = 8,
  NfeatPerClus = 10,
  returnDF = F,
  title = "",
  base_size = 12
)
}
\arguments{
\item{plotDF}{A data frame of numeric values to plot.}

\item{centers}{Number of centers to use for k-means clustering.}

\item{NfeatPerClus}{Number of features to include in the plot per cluster.}

\item{returnDF}{Logical indicating if the plotted data frame should be returned.}

\item{title}{Character string for the plot title.}
}
\value{
If returnDF is `TRUE`, the plotted data frame is returned.
}
\description{
Plot 2D dimensional reduction of a data frame.
}
