% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Seurat_based.R
\name{plot3d_seurat}
\alias{plot3d_seurat}
\title{Plot 3D visualization of Seurat object}
\usage{
plot3d_seurat(
  SerObj,
  feature1,
  feature2,
  feature3,
  color_feature = NULL,
  cols = colorRamp(c("navy", "red")),
  plot_title = "",
  slot = "data"
)
}
\arguments{
\item{SerObj}{A Seurat object}

\item{feature1}{The name of the first feature to plot along the x-axis}

\item{feature2}{The name of the second feature to plot along the y-axis}

\item{feature3}{The name of the third feature to plot along the z-axis}

\item{color_feature}{The name of the feature to use for coloring the points (optional)}

\item{cols}{The color palette to use for the points (optional)}

\item{plot_title}{The title of the plot (optional)}
}
\value{
A 3D plotly visualization of the Seurat object
}
\description{
Plot 3D visualization of Seurat object
}
