% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{plot_gene_network}
\alias{plot_gene_network}
\title{Make a GIF from plot list}
\usage{
plot_gene_network(count_matrix, method = "dist")
}
\arguments{
\item{count_matrix}{a count matrix}

\item{method}{dist or cor}
}
\value{
plots a network graph
}
\description{
This function takes in a list of ggplots and makes a GIF
}
