% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Network.R
\name{plot_knn}
\alias{plot_knn}
\title{Plot k-Nearest Neighbors Graph}
\usage{
plot_knn(df, k = 1, labels = NULL)
}
\arguments{
\item{df}{A dataframe with x and y coordinates and a label column defining the classes.}

\item{k}{An integer that defines the number of nearest neighbors to consider. Default is 1.}

\item{labels}{An optional character vector specifying the label column in the dataframe. 
If not specified, it is assumed that the label column is named "label".}
}
\value{
A plot of the k-nearest neighbors graph using ggplot2.
}
\description{
This function takes a dataframe that contains x and y coordinates and a label
that defines the classes and makes a plot of the k-nearest neighbors graph using 
Euclidean distance and the cover tree algorithm.
}
\examples{
# create sample data
set.seed(123)
x <- rnorm(100)
y <- rnorm(100)
label <- sample(c("A", "B", "C"), 100, replace = TRUE)
df <- data.frame(x = x, y = y, label = label)

# plot with k = 2
plot_knn(df, k = 2)

}
