% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Network.R
\name{plot_knn_labels}
\alias{plot_knn_labels}
\title{Plot k-Nearest Neighbors Graph with multiple lables}
\usage{
plot_knn_labels(
  df,
  k = 1,
  alpha = 0.7,
  algo = c("kd_tree", "cover_tree", "CR", "brute")
)
}
\arguments{
\item{df}{A dataframe with x and y coordinates and a label column defining the classes.}

\item{k}{An integer that defines the number of nearest neighbors to consider. Default is 1.}
}
\value{
A plot of the k-nearest neighbors graph using ggplot2.
}
\description{
This function takes a dataframe that contains x and y coordinates and a label
that defines the classes and makes a plot of the k-nearest neighbors graph using 
Euclidean distance and the cover tree algorithm.
}
