% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Seurat_based.R
\name{plot_libSize}
\alias{plot_libSize}
\title{plot_libSize}
\usage{
plot_libSize(
  SerObj,
  assay = "ADT",
  slot = "counts",
  featureSplit = NULL,
  col_vector = NULL
)
}
\arguments{
\item{SerObj}{A Seurat Object}

\item{assay}{Assay within the seurat object}

\item{slot}{which ADT slot data or count}

\item{featureSplit}{a feature of metadata; if NULL no split id done.}
}
\value{
A ggplot object
}
\description{
Plots a histogram of ADT/protein/CITE-seq markers (optional: identify percent of cells cut by a threshold)
}
