% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Seurat_based.R
\name{plot_loadings_coordinates.seurat}
\alias{plot_loadings_coordinates.seurat}
\title{This function generates a plot of gene loadings along genomic coordinates based on a Seurat object.}
\usage{
plot_loadings_coordinates.seurat(
  SerObj,
  reduction = "pca",
  redLab = "PC",
  loadingsLS = NULL,
  highlight_genes = NULL,
  TopNpos = 10,
  TopNneg = 10,
  data_set = "hsapiens_gene_ensembl",
  dimN = 1,
  invertWeights = F,
  includeUnMapped = T,
  geneLocPath = NULL
)
}
\arguments{
\item{SerObj}{A Seurat object.}

\item{reduction}{The dimensionality reduction method used in Seurat (default is "pca").}

\item{loadingsLS}{Give a list of loadings to use instead of Seurat object SDA reduction.}

\item{highlight_genes}{A character vector of gene symbols to highlight in the plot (default is NULL).}

\item{TopNpos}{The number of top positive loadings to display (default is 10).}

\item{TopNneg}{The number of top negative loadings to display (default is 10).}

\item{data_set}{passed to SDAtools:::get.location data_set default "hsapiens_gene_ensembl" can be  #mmulatta_gene_ensembl .. see ??useMart or do: mart = useMart('ensembl'), followed by listDatasets(mart).}

\item{dimN}{the reduction dim default = 1}

\item{invertWeights}{to invert the loading weight i.e., -loadings}

\item{includeUnMapped}{to include un mapped genes e.g. LOCs default = T}
}
\value{
A ggplot object representing the loadings along genomic coordinates.
}
\description{
This function generates a plot of gene loadings along genomic coordinates based on a Seurat object.
}
