% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Network.R
\name{plot_minium_spanning_tree}
\alias{plot_minium_spanning_tree}
\title{Plot Minimum Spanning Tree}
\usage{
plot_minium_spanning_tree(df, labels = NULL)
}
\arguments{
\item{df}{A data frame with x and y coordinates.}

\item{labels}{A vector of labels for each observation in \code{df}.}
}
\value{
A plot of the minimum spanning tree.
}
\description{
Plots the minimum spanning tree of a given dataset.
}
\examples{

# create sample data
set.seed(123)
x <- rnorm(100)
y <- rnorm(100)
label <- sample(c("A", "B", "C"), 100, replace = TRUE)
df <- data.frame(x = x, y = y, label = label)

# plot the minimum spanning tree
plot_minimum_spanning_tree(df)

}
