% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Seurat_based.R
\name{plot_pca_variance}
\alias{plot_pca_variance}
\title{Plot Explained Variance of Principal Components}
\usage{
plot_pca_variance(SerObj, pcs = 1:10)
}
\arguments{
\item{SerObj}{A Seurat object.}

\item{pcs}{Numeric vector specifying which PCs to plot. Default is 1:10.}
}
\value{
A ggplot object representing the explained variance of the specified principal components.
}
\description{
This function generates a bar plot illustrating the explained variance of principal components (PCs) in a Seurat object.
}
