% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Plot_Pseudotime_V_Gene}
\alias{Plot_Pseudotime_V_Gene}
\title{Plot a scatter plot with smoothed lines for selected features}
\usage{
Plot_Pseudotime_V_Gene(
  SerObj,
  SerObjrtByName,
  Feats = NULL,
  base_size = 20,
  col_vector = col_vector,
  showScatter = T,
  dowsampleScatter = T,
  scatterAlpha = 0.5
)
}
\arguments{
\item{SerObj}{A Seurat object to plot.}

\item{SerObjrtByName}{The name of the variable to use for SerObjrting cells on the x-axis.}

\item{Feats}{A vector of feature names to plot on the y-axis. Default is NULL, which will plot all features in the object.}

\item{base_size}{Base size of points and lines. Default is 20.}

\item{col_vector}{color vector}

\item{showScatter}{boolean default T to show scatter points}

\item{dowsampleScatter}{boolean default T to downsample the scatter points}

\item{scatterAlpha}{alpha value default 0.5 to show scatter points}
}
\value{
A ggplot object.
}
\description{
This function creates a scatter plot of a Seurat object, with smoothed lines for each selected feature. The x-axis is defined by the selected SerObjrtByName parameter, and the y-axis is defined by the values of the selected features added to the data frame. The function uses ggplot2 for visualization.
}
\examples{
Plot_Pseudotime_V_Gene(SerObj, "nCount_RNA", c("TNFRSF4", "CD69"))

}
