% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Plot_Pseudotime_V_Gene_combo}
\alias{Plot_Pseudotime_V_Gene_combo}
\title{Plot Violin Plots of Gene Expression by Group combo version}
\usage{
Plot_Pseudotime_V_Gene_combo(
  seuratObj,
  SortByName,
  Feats_pos = NULL,
  Feats_neg = NULL,
  base_size = 20,
  col_vector = col_vector,
  showScatter = TRUE,
  downsampleScatter = TRUE,
  scatterAlpha = 0.5,
  getWindows = T,
  decreasing = F,
  group.by = "Population"
)
}
\arguments{
\item{seuratObj}{A Seurat object containing the data to plot.}

\item{Feats_pos}{The name of the feature to plot from pos side}

\item{Feats_neg}{The name of the feature to plot from neg side}

\item{getWindows}{Bool default T to get bins to split}

\item{decreasing}{prder of pseudotime .}

\item{group.by}{The name of the metadata field to group cells by.}

\item{NumFeatName}{The name of the variable containing the number of features for each cell. If provided, cells with a number of features greater than the median will be excluded from the plot.}

\item{cutGT}{If TRUE (default), cells with a number of features greater than the median will be excluded from the plot. If FALSE, cells with a number of features less than or equal to the median will be excluded from the plot.}

\item{ThrCut}{The threshold for feature filtering. Cells with a value in the sdaCompName column greater than this threshold will be excluded from the plot. Default is 0, which will skip this step.}

\item{GrpFacLevels}{The levels of the factor variable to use for grouping. If NULL (default), all levels will be included.}

\item{comparisons}{A list of comparisons to perform using \code{stat_compare_means}. Each comparison should be a vector of two group names.}

\item{xlab}{(Optional) The x-axis label.}

\item{ylab}{(Optional) The y-axis label.}

\item{palette}{The color palette to use for the plot.}

\item{addJitter}{If TRUE, jitter points will be added to the plot. Default is FALSE.}
}
\value{
A ggplot object.
}
\description{
Plot Violin Plots of Gene Expression by Group combo version
}
