% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot a scatter plot with smoothed lines for selected features}
\usage{
plot_scatter(seuratObj, SortByName, Feats = NULL, base_size = 20)
}
\arguments{
\item{seuratObj}{A Seurat object to plot.}

\item{SortByName}{The name of the variable to use for sorting cells on the x-axis.}

\item{Feats}{A vector of feature names to plot on the y-axis. Default is NULL, which will plot all features in the object.}

\item{base_size}{Base size of points and lines. Default is 20.}
}
\value{
A ggplot object.
}
\description{
This function creates a scatter plot of a Seurat object, with smoothed lines for each selected feature. The x-axis is defined by the selected SortByName parameter, and the y-axis is defined by the values of the selected features added to the data frame. The function uses ggplot2 for visualization.
}
\examples{
plot_scatter(seuratObj, "nCount_RNA", c("TNFRSF4", "CD69"))

}
