% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Seurat_based.R
\name{plot_violin_wpvalue}
\alias{plot_violin_wpvalue}
\title{Plot violin plots with Wilcoxon p-values}
\usage{
plot_violin_wpvalue(
  SerObj,
  feature,
  group.by = "ident",
  layer = "data",
  GrpFacLevels = NULL,
  assay = "RNA",
  colors = col_vector,
  comparisons = NULL,
  title = "",
  doJitter = T
)
}
\arguments{
\item{SerObj}{A Seurat object containing the data.}

\item{feature}{The feature to plot.}

\item{group.by}{The grouping variable to use for the plot (default: "ident").}

\item{layer}{The layer to use for the plot (default: "data").}

\item{GrpFacLevels}{Optional vector of grouping factor levels to plot.}

\item{assay}{The assay to use for the plot (default: "RNA").}

\item{colors}{A vector of colors to use for the plot.}

\item{comparisons}{A list of pairwise comparisons to perform.}

\item{title}{A string to use as the plot title.}

\item{doJitter}{Whether to add jitter to the points (default: TRUE).}
}
\value{
A ggplot object.
}
\description{
This function generates a plot of violin plots with Wilcoxon p-values
for the given feature and groupings in a Seurat object.
}
