% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Plot a word cloud from a named vector of word counts}
\usage{
plot_wordcloud(
  word_count,
  max.words = 100,
  colors = brewer.pal(8, "Dark2"),
  min.freq = 1
)
}
\arguments{
\item{word_count}{A named vector where the names are words and the values are counts}

\item{max.words}{The maximum number of words to display in the word cloud (default: 100)}

\item{colors}{The color palette to use for the word cloud (default: brewer.pal(8, "Dark2"))}

\item{min.freq}{The minimum frequency of a word to be included in the word cloud (default: 1)}
}
\value{
A plot of the word cloud
}
\description{
Plot a word cloud from a named vector of word counts
}
