% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot Enrichment}
\usage{
plotEnrichment(
  SerObj,
  field1,
  field2,
  size,
  fontsize,
  titlefontsize,
  fullorder
)
}
\arguments{
\item{SerObj}{A Seurat object.}

\item{field1}{Field 1 to compare.}

\item{field2}{Field 2 to compare.}

\item{size}{Size of plot.}

\item{fontsize}{Font size for labels.}

\item{titlefontsize}{Font size for plot title.}

\item{fullorder}{}
}
\value{
A plot of enrichment analysis between two fields.
}
\description{
Plot Enrichment
}
