% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{Plot Gene Expression in Seurat Object}
\usage{
plotGeneExpression(
  serObj,
  gene,
  threshold = 0.5,
  groupBy = "RNA_snn_res.0.4",
  doPie = FALSE,
  col_vector = col_vector
)
}
\arguments{
\item{gene}{The name of the gene for which expression is to be plotted.}

\item{threshold}{The expression threshold for selecting cells (default is 0.5).}

\item{groupBy}{The variable by which the data is grouped (default is "RNA_snn_res.0.4").}

\item{doPie}{Logical, if TRUE, a pie chart is generated instead of a violin plot (default is FALSE).}

\item{SerObj}{A Seurat object containing single-cell RNA-seq data.}
}
\description{
This function generates a violin plot or pie chart for the expression of a specified gene
in a Seurat object, based on a given threshold.
}
