% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{plotPCgeneLoadings}
\alias{plotPCgeneLoadings}
\title{scatter-bubble style of selected PCs and genes}
\usage{
plotPCgeneLoadings(SerObj, features, pcs = c(1, 2), quaT = 0.8, base_size = 10)
}
\arguments{
\item{SerObj}{a seurat obj}

\item{features}{features to show}

\item{pcs}{a numeric vector of 2 pc nnumbers e.g. c(1, 2)}

\item{quaT}{color pink via quantiles}

\item{base_size}{base_size of plot}
}
\value{
ggplot scatter-bubble plot style
}
\description{
scatter-bubble style of selected PCs and genes
}
