% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_util.R
\name{predict.SDA}
\alias{predict.SDA}
\title{Predict using SDA}
\usage{
predict.SDA(new.data, sda.model, return_cell_score = T)
}
\arguments{
\item{new.data}{a matrix containing new data to be predicted}

\item{sda.model}{a fitted sparse discriminant analysis model}

\item{return_cell_score}{a logical indicating whether to return the cell score or the predicted new data}
}
\value{
If return_cell_score is T, returns cell score, else returns the predicted new data.
}
\description{
Predict using sparse discriminant analysis
}
