% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_Seurat.R
\name{prepNormedDGE_SDA}
\alias{prepNormedDGE_SDA}
\title{Prepare and Save Normalized Gene Expression Data for SDA}
\usage{
prepNormedDGE_SDA(normedDGE, outputFolder)
}
\arguments{
\item{normedDGE}{A matrix or data frame containing normalized gene expression data,
ready for analysis. This data is expected to be preprocessed and normalized appropriately.}

\item{outputFolder}{A character string specifying the path to the output folder where
the normalized data and results will be saved. The function will create the directory
if it does not exist, and will ensure it ends with a slash for consistency in path handling.}
}
\value{
A list containing two elements: `resultsDir` indicating the path to the results
  directory, and `rawDataFile` specifying the path to the saved normalized gene expression data file.
}
\description{
This function prepares the environment for single-cell data analysis (SDA) by saving
normalized gene expression data to a specified output folder. It ensures that the
output directory exists, appending a trailing slash if necessary, and utilizes
SDAtools for data export. The function also prepares a results directory within
the output folder for further analysis.
}
