% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{princurve_bootstrap}
\alias{princurve_bootstrap}
\title{Principal Curve Bootstrap}
\usage{
princurve_bootstrap(
  input,
  df_range = NA,
  df_start = NA,
  df_end = NA,
  plot = TRUE
)
}
\arguments{
\item{input}{A matrix or data frame representing the input data.}

\item{df_range}{A numeric vector specifying the range of degrees of freedom (df) for principal curve fitting.
If any element is set to NA, the range will be determined by df_start and df_end parameters.}

\item{df_start}{A numeric value indicating the starting degree of freedom for principal curve fitting.
Ignored if df_range is specified.}

\item{df_end}{A numeric value indicating the ending degree of freedom for principal curve fitting.
Ignored if df_range is specified.}

\item{plot}{A logical value indicating whether to plot the principal curves.}
}
\value{
A list containing the principal curves fitted for different degrees of freedom.
}
\description{
This function performs bootstrap estimation of principal curves based on the given input data.
}
