% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Spatial_based.R, R/Seurat_based.R
\name{process_SCTbase}
\alias{process_SCTbase}
\title{Process single cell transcriptomics data with SCT transformation}
\usage{
process_SCTbase(
  serObj = NULL,
  ncells = 3000,
  assay = "Spatial",
  dims = 1:30,
  verbose = T,
  clusterResolutions = c(0.2, 0.4, 0.6, 0.8, 1.2)
)

process_SCTbase(
  serObj = NULL,
  ncells = 3000,
  assay = "Spatial",
  dims = 1:30,
  verbose = T,
  clusterResolutions = c(0.2, 0.4, 0.6, 0.8, 1.2)
)
}
\arguments{
\item{serObj}{a seurat obj}

\item{ncells}{The number of cells used to build NB regression, default is 3000}

\item{assay}{The assay used to run the SCT transformation, default is "Spatial"}

\item{dims}{The dimensions to use for running PCA, t-SNE, UMAP, default is 1:30}

\item{verbose}{Logical, whether to print progress messages, default is TRUE}

\item{clusterResolutions}{A numeric vector of resolutions to use for finding clusters, default is c(0.2, 0.4, 0.6, 0.8, 1.2)}
}
\value{
A seurat object with added dimensionality reductions, clustering, and cell type annotations.

A Processed Seurat Obj
}
\description{
This function takes in a seurat object and applies various transformations and clustering algorithms. The end goal is to identify cell types and subpopulations within the data.

scatter-bubble style of selected PCs and genes
}
