% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOSAP_based.R
\name{ProcessSeurat1b}
\alias{ProcessSeurat1b}
\title{Run a variation (1b) of the primary seurat processing steps.}
\usage{
ProcessSeurat1b(
  seuratObj,
  saveFile = NULL,
  doCellCycle = T,
  doCellFilter = F,
  nCount_RNA.high = 20000,
  nFeature.high = 3000,
  pMito.high = 0.15,
  nCount_RNA.low = 0.99,
  nFeature.low = 200,
  pMito.low = -Inf,
  forceReCalc = F,
  variableGeneTable = NULL,
  variableFeatureSelectionMethod = "vst",
  nVariableFeatures = 2000,
  printDefaultPlots = T,
  npcs = 50,
  ccPcaResultFile = NULL,
  useSCTransform = F,
  mean.cutoff = c(0.0125, 3),
  dispersion.cutoff = c(0.5, Inf),
  spikeGenes = NULL,
  excludedGenes = NULL,
  verbose = FALSE,
  Regress_basic = F
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{saveFile}{If provided, the seuratObj will be saved here as it is processed, providing some ability to resume if there is a failure}

\item{doCellCycle}{If true, CellCycle genes will be regressed}

\item{doCellFilter}{If true, basic filtering will be performed using nCount_RNA, nFeature_RNA, and pMito}

\item{nCount_RNA.high}{If doCellFilter=T, cells with nCount_RNA above this value will be filtered}

\item{nFeature.high}{If doCellFilter=T, cells with nFeature above this value will be filtered}

\item{pMito.high}{If doCellFilter=T, cells with percent mito above this value will be filtered}

\item{nCount_RNA.low}{If doCellFilter=T, cells with nCount_RNA below this value will be filtered}

\item{nFeature.low}{If doCellFilter=T, cells with nFeature below this value will be filtered}

\item{pMito.low}{If doCellFilter=T, cells with percent mito  below this value will be filtered}

\item{forceReCalc}{If true, all steps will be repeated even if already marked as complete}

\item{variableGeneTable}{If provided, a table of variable genes will be written to this file}

\item{variableFeatureSelectionMethod}{The selection method to be passed to FindVariableFeatures()}

\item{nVariableFeatures}{The number of variable features to find}

\item{printDefaultPlots}{If true, the default set of QC plots will be printed}

\item{npcs}{Number of PCs to use for RunPCA()}

\item{ccPcaResultFile}{If provided, the PCA results from cell cycle regression will be written here}

\item{useSCTransform}{If true, SCTransform will be used in place of the standard Seurat workflow (NormalizeData, ScaleData, FindVariableFeatures)}

\item{mean.cutoff}{Passed directly to FindVariableFeatures}

\item{dispersion.cutoff}{Passed directly to FindVariableFeatures}

\item{spikeGenes}{If provided these will be appended to the set of VariableFeatures}

\item{excludedGenes}{If provided these will be removed from the set of VariableFeatures}

\item{Regress_basic}{if true p.mito and nCount_RNA are used to do a basic regression to remove the associated 'noise'. By default, this is F because by nature of the method, we are biasing the data.}
}
\value{
A modified Seurat object.
}
\description{
This is the 1b entry point for processing (v1b) scRNAseq data with Seurat
}
