% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{ProduceComboHeatmap.old}
\alias{ProduceComboHeatmap.old}
\title{Produce a combination heatmap old ver}
\usage{
ProduceComboHeatmap.old(
  seuratObj,
  markerVec,
  pairedList,
  pairedList2 = NULL,
  labelColumn,
  prefix,
  adtoutput = "unpaired",
  rowsplit = NULL,
  columnsplit = NULL,
  size,
  coldend = TRUE,
  rowdend = TRUE,
  coldendside = "bottom",
  rowdendside = "left",
  row_names_side = "left",
  column_names_side = "bottom",
  fontsize = 12,
  titlefontsize = 20,
  gap = 0,
  show_column_dend = T,
  show_row_dend = T
)
}
\arguments{
\item{seuratObj}{A seurat object}

\item{markerVec}{A vector of markers}

\item{pairedList}{A list of paired data}

\item{pairedList2}{A second list of paired data}

\item{labelColumn}{Column name for sample labels}

\item{prefix}{Prefix for output file names}

\item{adtoutput}{Output type (default: "unpaired")}

\item{rowsplit}{Split data by row (default: NULL)}

\item{columnsplit}{Split data by column (default: NULL)}

\item{size}{Plot size}

\item{coldend}{Show dendrogram for columns (default: TRUE)}

\item{rowdend}{Show dendrogram for rows (default: TRUE)}

\item{coldendside}{Side to place column dendrogram (default: "bottom")}

\item{rowdendside}{Side to place row dendrogram (default: "left")}

\item{fontsize}{Font size (default: 12)}

\item{titlefontsize}{Title font size (default: 20)}

\item{gap}{Gap between panels (default: 0)}
}
\value{
A list of differentially expressed genes
}
\description{
Produce a combination heatmap old ver
}
