% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{ProduceComboHeatmapFromMatrix}
\alias{ProduceComboHeatmapFromMatrix}
\title{Produce Combo Heatmap From Matrix}
\usage{
ProduceComboHeatmapFromMatrix(
  SerObj,
  mat,
  markerVec,
  pairedList,
  pairedList2,
  labelColumn,
  prefix,
  adtoutput = "unpaired",
  rowsplit = NULL,
  columnsplit = NULL,
  size,
  coldend = TRUE,
  rowdend = TRUE,
  coldendside = "bottom",
  rowdendside = "left",
  fontsize = 12,
  titlefontsize = 14,
  gap = 0
)
}
\arguments{
\item{SerObj}{A Seurat object containing the single-cell RNA-seq data.}

\item{mat}{A matrix object containing the data for generating the heatmap.}

\item{markerVec}{A vector of marker genes to be used for generating the heatmap.}

\item{pairedList}{A list of paired conditions for the first set of markers.}

\item{pairedList2}{A list of paired conditions for the second set of markers.}

\item{labelColumn}{A column name or index in the metadata of SerObj to be used as labels for the heatmap.}

\item{prefix}{A prefix for the output file name.}

\item{adtoutput}{Output type for ADT features. Default is "unpaired".}

\item{rowsplit}{A vector of row splits for the heatmap.}

\item{columnsplit}{A vector of column splits for the heatmap.}

\item{size}{Size of the heatmap. Default is NULL.}

\item{coldend}{A logical value indicating whether to add dendrogram on the columns. Default is TRUE.}

\item{rowdend}{A logical value indicating whether to add dendrogram on the rows. Default is TRUE.}

\item{coldendside}{A character value specifying the side to place the column dendrogram. Default is "bottom".}

\item{rowdendside}{A character value specifying the side to place the row dendrogram. Default is "left".}

\item{fontsize}{Font size for the heatmap labels. Default is 12.}

\item{titlefontsize}{Font size for the title of the heatmap. Default is 14.}

\item{gap}{Gap between the heatmaps. Default is 0.}
}
\value{
A combo heatmap plot.
}
\description{
This function produces a combination heatmap from a matrix using Seurat object.
}
