% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocle.R
\name{PseudoMySeur}
\alias{PseudoMySeur}
\title{PseudoMySeur}
\usage{
PseudoMySeur(SerObj, features = NULL, removeCells = NULL, removeNoisyFeats = T)
}
\arguments{
\item{SerObj}{A Seurat object.}

\item{features}{A character vector specifying the features (genes) to include in the analysis. If NULL, all features are used.}

\item{removeCells}{A character vector specifying the cell names to remove from the analysis. If NULL, no cells are removed.}

\item{removeNoisyFeats}{A logical value indicating whether to remove noisy features. Default is TRUE.}
}
\value{
A Monocle CellDataSet object representing the pseudo-time trajectory.
}
\description{
A function to create a pseudo-time trajectory using Monocle on a Seurat object.
}
