% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{quantile_breaks}
\alias{quantile_breaks}
\title{Quantile Breaks}
\usage{
quantile_breaks(xs, n = 10)
}
\arguments{
\item{xs}{A numeric vector of values.}

\item{n}{An integer specifying the number of quantile breaks to be calculated.}
}
\value{
A numeric vector representing the quantile breaks.
}
\description{
Quantile Breaks
}
\examples{
xs <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
quantile_breaks(xs, n = 5)

}
