% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalizations.R
\name{quantile_norm}
\alias{quantile_norm}
\title{quantile_norm}
\usage{
quantile_norm(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A quantile-based normalized matrix
}
\description{
quantile normalize a matrix
}
