% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Seurat_based.R
\name{QuantileADTcounter}
\alias{QuantileADTcounter}
\title{QuantileADTcounter}
\usage{
QuantileADTcounter(
  so,
  features = NULL,
  min.cells = 10,
  slot = "data",
  thresh = NULL
)
}
\arguments{
\item{so}{A Seurat Object}

\item{features}{Features}

\item{min.cells}{a horizontal line is plotted at this value}

\item{slot}{which ADT slot data or count}

\item{thresh}{a vertical line is plotted at this value and perc. of cells below it is computed and plotted}
}
\value{
A ggplot object
}
\description{
Plots a histogram of ADT/protein/CITE-seq markers optional: identify percent of cells cut by a threshold
}
