% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{QuantileEdgeClean}
\alias{QuantileEdgeClean}
\title{QuantileEdgeClean}
\usage{
QuantileEdgeClean(
  mat,
  LowSet = 0,
  HighSet = NULL,
  lowQuant = 0.1,
  highQuant = 0.9
)
}
\arguments{
\item{mat}{a numerical matrix, rows are features cols are cells/measured units}

\item{LowSet}{a numerical value for lower send quantiles; default 0}

\item{HighSet}{a numerical value for higher send quantiles; default NULL sets to high quantile}

\item{lowQuant}{a value to set the low quantiles}

\item{highQuant}{a value to set the high quantiles}
}
\value{
A numeric mat
}
\description{
Scales the range of input to be between 0 and 1
}
