% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale.R, R/Transformations.R
\name{range01}
\alias{range01}
\title{Convert Vector to 0-1 Range}
\usage{
range01(x, MaxN = NULL, MinN = NULL)

range01(x, MaxN = NULL, MinN = NULL)
}
\arguments{
\item{x}{a numerical vector}

\item{MaxN}{set Max value; NULL default find max(x)}

\item{MinN}{set Min value; NULL default find min(x)}
}
\value{
Numeric vector scaled to range of 0 to 1.

A numeric vector
}
\description{
This function converts a numeric vector to a range of 0 to 1 by linearly scaling the values
between the minimum and maximum of the input vector.

Scales the range of input to be between 0 and 1
}
\examples{
x <- c(2, 5, 9, 11, 14)
range01(x)
}
\keyword{range}
