% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Rank_Swarm}
\alias{Rank_Swarm}
\title{Rank a feature and plot a swarm grouped by select categorical metadata}
\usage{
Rank_Swarm(
  SerObj,
  RankFeat,
  group.by,
  title = "Cell SerObjrted",
  color.by = NULL
)
}
\arguments{
\item{SerObj}{A Seurat object to perform gating on.}

\item{RankFeat}{A feature to rank, like SDA comp}

\item{group.by}{A categorical meta feature}

\item{title}{A title}

\item{color.by}{A categorical meta feature default NULL to color the point}
}
\value{
ggplot viz of ranking per group
}
\description{
Rank a feature and plot a swarm grouped by select categorical metadata
}
