% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{RearrangeLabelsByFrequency}
\alias{RearrangeLabelsByFrequency}
\title{RearrangeLabelsByFrequency}
\usage{
RearrangeLabelsByFrequency(vec)
}
\value{
a vector relabeled frequency wise
}
\description{
reassign the labels of the elements in the vector based on the frequency of their original labels.
}
