% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{ReduceSerObj_HTO}
\alias{ReduceSerObj_HTO}
\title{ReduceSerObj_HTO}
\usage{
ReduceSerObj_HTO(SerObj, removeHTOs = NULL)
}
\arguments{
\item{SerObj}{A Seurat Object}

\item{removeHTOs}{Vector of names of HTOs to remove, if NULL, removeHTOs = c("Discordant", "Doublet", "ND")}
}
\value{
A subset Seurat object.
}
\description{
Remove unwanted HTOS from Seurat objects
}
