% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rotate_df}
\alias{rotate_df}
\title{rotate_df}
\usage{
rotate_df(df, angle)
}
\arguments{
\item{angle, }{rotation angle}

\item{tsne, }{a 2D df scatter sctructure}
}
\value{
rotated tsne Y
}
\description{
rotates a 2D df like tsne UMAP etc any scatter
}
