% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{rotate_tsne}
\alias{rotate_tsne}
\title{rotate_tsne}
\usage{
rotate_tsne(tsne, angle)
}
\arguments{
\item{tsne, }{a tsne object with tsne$Y}

\item{angle, }{rotation angle}
}
\value{
rotated tsne Y
}
\description{
rotates a tsne object
}
