% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Run_Integration}
\alias{Run_Integration}
\title{Run Integration Pipeline for Single-Cell Data}
\usage{
Run_Integration(
  SerObj,
  do.UnIntg = T,
  do.Harmony = T,
  useVarGenes = TRUE,
  nfeatures = 2000,
  spikeInGenes = NULL,
  dims = 1:15,
  harmony_res = 0.2,
  rpca_res = 0.2,
  rpca_ref_res = 0.4,
  reference = 1,
  LigerK = 20,
  nIteration = 30,
  Liger_res = 0.4
)
}
\arguments{
\item{SerObj}{A Seurat object containing the single-cell RNA-seq data to be integrated.}

\item{do.UnIntg}{Logical. If `TRUE`, performs the unintegrated analysis using PCA. Default is `TRUE`.}

\item{do.Harmony}{Logical. If `TRUE`, performs the Harmony-based integration. Default is `TRUE`.}

\item{useVarGenes}{Logical. If `TRUE`, uses variable genes for integration. Default is `TRUE`.}

\item{nfeatures}{Numeric. The number of variable features to select if `useVarGenes` is `TRUE`. Default is `2000`.}

\item{spikeInGenes}{Character vector. A vector of spike-in genes to be added to the variable genes for semi-supervised integration. Default is `NULL`.}

\item{dims}{Numeric vector. Dimensions to use for PCA/UMAP and clustering (default is 1:15).}

\item{harmony_res}{Numeric. The resolution for clustering after Harmony integration. Default is `0.2`.}

\item{rpca_res}{Numeric. The resolution for clustering after RPCA integration. Default is `0.2`.}

\item{rpca_ref_res}{Numeric. The resolution for clustering after reference-based RPCA integration. Default is `0.4`.}

\item{reference}{Integer. Specifies which dataset to use as a reference for reference-based RPCA. Default is `1`.}

\item{LigerK}{Integer. The number of factors for LIGER integration. Default is `20`.}

\item{nIteration}{Integer. The number of iterations for LIGER's INMF algorithm. Default is `30`.}

\item{Liger_res}{Numeric. The resolution for clustering after LIGER integration. Default is `0.4`.}

\item{do.rPCA}{Logical. If `TRUE`, performs RPCA integration. Default is `TRUE`.}

\item{do.rPCAref}{Logical. If `TRUE`, performs reference-based RPCA integration. Default is `TRUE`.}

\item{do.LIGER}{Logical. If `TRUE`, performs LIGER-based integration. Default is `TRUE`.}
}
\value{
The Seurat object (`SerObj`) with added integration results, UMAP embeddings, and cluster assignments for each method run.
}
\description{
This function performs multiple integration methods (Unintegrated, Harmony, RPCA, reference-based RPCA, LIGER) on a Seurat object and returns the integrated object with clustering and UMAP embeddings.
}
\examples{
# Run integration with default settings
integrated_obj <- Run_Integration(SerObj)

# Run only Harmony and RPCA integration
integrated_obj <- Run_Integration(SerObj, do.UnIntg = FALSE, do.Harmony = TRUE, do.rPCA = TRUE, do.LIGER = FALSE)

}
