% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{RunComBat}
\alias{RunComBat}
\title{RunComBat}
\usage{
RunComBat(
  SerObj = NULL,
  batch = "batch",
  runningTime = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SerObj}{A seurat SerObj to correct batch effects.}

\item{batch}{Batch labels.}

\item{runningTime}{Return the running time.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
Seurat SerObj with the corrected data in the 'integrated' assay.
}
\description{
RunComBat
}
