% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{RunLiger}
\alias{RunLiger}
\title{RunLiger}
\usage{
RunLiger(
  SerObj = NULL,
  batch = "batch",
  k = 30,
  runningTime = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SerObj}{A seurat SerObj to correct batch effects.}

\item{batch}{Batch labels.}

\item{k}{Inner dimension of factorization (number of factors)}

\item{runningTime}{Return the running time.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
Seurat SerObj with the corrected data in the 'Liger' reduction.
}
\description{
RunLiger
}
