% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{RunScMerge}
\alias{RunScMerge}
\title{RunScMerge}
\usage{
RunScMerge(
  SerObj = NULL,
  batch = "batch",
  ks = NULL,
  runningTime = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SerObj}{A seurat SerObj to correct batch effects.}

\item{batch}{Batch labels.}

\item{ks}{A vector indicates the kmeans's K for each batch, which length needs to be the same as the number of batches.}

\item{runningTime}{Return the running time.}

\item{verbose}{Print verbose.}

\item{...}{Arguments passed to other methods.}
}
\value{
Corrected Seurat SerObj.
}
\description{
RunScMerge
}
