% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{RuntSNE.Matrix}
\alias{RuntSNE.Matrix}
\title{This function to run tsne on matrix}
\usage{
RuntSNE.Matrix(
  myMat,
  save_path = NULL,
  tsnepp,
  tSNE_n.iter,
  num_threads = 8,
  check_duplicates = F,
  selectFeats = NULL,
  overwrite = F
)
}
\arguments{
\item{myMat}{SDA score matrix where cols are the comps and rows are cells}

\item{save_path}{default NULL ignores saving, or set path to save the tsne rds obj}

\item{tsnepp}{tSNE preplexity}

\item{tSNE_n.iter}{N iter for tSNE}

\item{num_threads}{number of threads}

\item{check_duplicates}{check for dup rows}

\item{selectFeats}{select Features from Matrix to run the tSNE}
}
\value{
A list of two vector text of enrichments, one for pos one for neg scored cells
}
\description{
This function to run tsne on matrix
}
