% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{save_ggplot_list}
\alias{save_ggplot_list}
\title{Save a patchwork of ggplot figures to a file}
\usage{
save_ggplot_list(
  ggLS,
  basepath,
  filename,
  splitN = 20,
  width = 12,
  height = 7,
  units = "in",
  dpi = 150
)
}
\arguments{
\item{ggLS}{A list of ggplot figures.}

\item{basepath}{The base path for the output file.}

\item{filename}{The name of the output file, not including any extension.}

\item{splitN}{The number of ggplots to include in each patchwork file (default is 20).}

\item{width}{The width of the output file in inches (default is 12).}

\item{height}{The height of the output file in inches (default is 7).}

\item{units}{The units of the output file (default is "in").}

\item{dpi}{The resolution of the output file in dots per inch (default is 150).}
}
\description{
This function takes a list of ggplot figures and saves them as a patchwork to a file specified by a base path and filename, split into multiple files if the list is longer than a given split N.
}
\examples{
## Save a list of ggplot figures to a file
my_ggplots <- list(ggplot(data = mtcars, aes(x = mpg, y = wt)) + geom_point(),
                   ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width)) + geom_point())
save_ggplot_list(ggLS = my_ggplots, basepath = "/path/to/file", filename = "my_ggplots")

}
