% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{Save_LS_Excel}
\alias{Save_LS_Excel}
\title{Save a List of DE Results to an Excel File}
\usage{
Save_LS_Excel(DFLS, file_path = "DE_results.xlsx")
}
\arguments{
\item{DFLS}{A named list of data frames}

\item{file_path}{File path for the output Excel file}
}
\value{
Saves an Excel file with multiple sheets.
}
\description{
Saves a list of  data frames into an Excel file, with each list item as a separate sheet.
}
\examples{
save_de_results_to_excel(DEgeneLS, "DE_results.xlsx")
}
