% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{Save_LS_ExcelV2}
\alias{Save_LS_ExcelV2}
\title{Save DE Results to Excel, Splitting by Up/Down Regulation}
\usage{
Save_LS_ExcelV2(
  DFLS,
  file_path = "DE_results.xlsx",
  logfc_col = "log2FoldChange"
)
}
\arguments{
\item{DFLS}{Named list of data frames containing DE results.}

\item{file_path}{File path to save the Excel file (default: "DE_results.xlsx").}

\item{logfc_col}{Name of the column containing log2 fold changes (default: "log2FoldChange").}
}
\value{
Saves an Excel file with organized sheets.
}
\description{
This function takes a named list of DESeq2 results data frames (`DEgeneLS`),
splits each into **Upregulated (log2FoldChange > 0) and Downregulated (log2FoldChange < 0)**
genes, and saves them in an Excel workbook, creating separate sheets for each.
}
\examples{
Save_LS_ExcelV2(DEgeneLS, file_path = "DE_results.xlsx")
}
