% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_df.R
\name{SaveMy_table}
\alias{SaveMy_table}
\title{Generate and Save/Display a Styled Table}
\usage{
SaveMy_table(df, file_name, save = TRUE)
}
\arguments{
\item{df}{A dataframe to be formatted into a table.}

\item{file_name}{A string specifying the name of the file if saving the table.}

\item{save}{A logical value indicating whether to save the table (`TRUE`) or just display it (`FALSE`). Default is `TRUE`.}
}
\value{
If `save = FALSE`, the table is printed in the console. If `save = TRUE`, it is saved to the specified file.
}
\description{
This function creates a styled table from a given dataframe and either displays it 
or saves it as an image.
}
\examples{
SaveMy_table(mtcars, "table.png", save = TRUE)
SaveMy_table(iris, "table.html", save = FALSE)
}
