% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.R
\docType{data}
\name{scale_color_gradient_heat2}
\alias{scale_color_gradient_heat2}
\title{scale_color_gradient_heat2}
\format{
An object of class \code{ScaleContinuous} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 18.
}
\usage{
scale_color_gradient_heat2
}
\value{
A vector of colors.
}
\description{
a vector of colors that are in sequence well for a heatmap
}
\keyword{datasets}
