% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_themes.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Scale Colour Publication}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the discrete_scale function.}
}
\description{
This function sets the color scale to a predefined set of colors commonly used for publication purposes.
}
