% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_themes.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Scale Fill Publication}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the discrete_scale function.}
}
\description{
This function sets the fill color scale to a predefined set of colors commonly used for publication purposes.
}
