% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{scatter_plot_seurat}
\alias{scatter_plot_seurat}
\title{scatter_plot_seurat}
\usage{
scatter_plot_seurat(
  seurat_obj,
  meta_feature,
  gene_name,
  x_split_value,
  y_split_value,
  assay = "RNA",
  layer = "counts",
  base_size = 11
)
}
\arguments{
\item{seurat_obj}{A Seurat Object}
}
\value{
ggplot
}
\description{
create scatter plot
}
