% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioStudies_ArrayExpress_based.R
\name{scrape_arrayexpress}
\alias{scrape_arrayexpress}
\title{Scrape ArrayExpress metadata and fastq.gz links}
\usage{
scrape_arrayexpress(
  accession,
  port = 4562L,
  browser = "firefox",
  LoadWaitSec = 5,
  R1R2only = T,
  splitData = F
)
}
\arguments{
\item{accession}{Character string specifying the ArrayExpress accession number.}
}
\value{
No direct return value. The function saves metadata as a CSV file and fastq.gz links as a text file.
}
\description{
This function scrapes metadata and fastq.gz links from ArrayExpress for a given accession.
}
\examples{
\dontrun{
# Example usage:
scrape_arrayexpress("E-MTAB-1234")
}
}
\seealso{
[rvest](https://cran.r-project.org/web/packages/rvest/index.html), [RSelenium](https://cran.r-project.org/web/packages/RSelenium/index.html)
}
