% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_processing_based.R
\name{SDA.AddCompStats}
\alias{SDA.AddCompStats}
\title{SDA.AddCompStats}
\usage{
SDA.AddCompStats(
  SDAres,
  sdThrsh = 0.04,
  maxscoreThrsh = 20,
  maxloadThrsh = 1,
  redoCalc = T
)
}
\arguments{
\item{SDAres}{An SDA results list/object}
}
\value{
An SDA results list/object with Comp statistics
}
\description{
SDA.AddCompStats
}
