% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_processing_based.R
\name{SDA.GO_Enrichment}
\alias{SDA.GO_Enrichment}
\title{Perform Gene Ontology Enrichment Analysis on SDA Components}
\usage{
SDA.GO_Enrichment(
  sdaResults,
  components,
  orgDb = "org.Hs.eg.db",
  geneNumber = 100,
  direction = "Both"
)
}
\arguments{
\item{sdaResults}{A list containing SDA results, expected to include a 'loadings' element
with components and gene loadings.}

\item{components}{A numeric vector indicating which components from the SDA results to analyze.}

\item{orgDb}{A character string specifying the Bioconductor organism annotation package
to use for the enrichment analysis. Default is "org.Hs.eg.db" for human. Other examples
include "org.Mm.eg.db" for mouse and "org.Mmu.eg.db" for rhesus macaque.}

\item{geneNumber}{An integer specifying the number of top genes to consider for enrichment
analysis in each component and direction. Default is 100.}

\item{direction}{A character string indicating the direction of gene loading to consider
for the analysis. Can be "Pos" for positive, "Neg" for negative, or "Both" for both directions.
Default is "Both".}
}
\value{
A list where each element corresponds to a component and direction analyzed, containing
  the results of the GO enrichment analysis. Results include enrichment scores, gene ratios,
  background ratios, and odds ratios among others.
}
\description{
This function conducts Gene Ontology (GO) enrichment analysis on specified components
from Single-cell Data Analysis (SDA) results. It supports analysis in positive, negative,
or both directions and allows specifying the number of top genes to consider. The function
utilizes `clusterProfiler` for the enrichment analysis and is capable of handling different
organism databases.
}
