% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_Seurat.R
\name{SDA_ser_preproc}
\alias{SDA_ser_preproc}
\title{Preprocess Seurat Object for Single-cell Data Analysis}
\usage{
SDA_ser_preproc(
  seuratObj,
  assayName = "RNA",
  Layer = "counts",
  minFeatureCount = 1,
  minCellsExpressingFeature = 0,
  perCellExpressionThreshold = 0,
  featureInclusionList = NULL,
  featureExclusionList = NULL,
  maxFeaturesDiscarded = 0.75,
  minLibrarySize = 0
)
}
\arguments{
\item{seuratObj}{A Seurat object containing single-cell RNA sequencing data.}

\item{assayName}{The name of the assay to use, default is "RNA".}

\item{Layer}{The data layer to use, default is "counts".}

\item{minFeatureCount}{Minimum count for a feature to be included, default is 1.}

\item{minCellsExpressingFeature}{Minimum number of cells that must express a feature, default is 0.}

\item{perCellExpressionThreshold}{Threshold for expression level per cell, default is 0.}

\item{featureInclusionList}{A list of features to forcibly include, default is NULL.}

\item{featureExclusionList}{A list of features to exclude, default is NULL.}

\item{maxFeaturesDiscarded}{The maximum proportion (0-1) or absolute number of features to discard, default is 0.75.}

\item{minLibrarySize}{The minimum library size required for a cell to be included, default is 0.}
}
\value{
A list containing the filtered assay data (`SerObj.DGE`), the list of features used (`featuresToUse`),
        and the minimum library size considered (`minLibrarySize`).
}
\description{
This function prepares a Seurat object for single-cell data analysis (SDA) by filtering features (genes)
and cells based on various criteria such as minimum feature count, minimum number of cells expressing a feature,
expression thresholds, inclusion and exclusion lists, and library size. It also checks and warns about 
the maximum number of cells the SDA can handle. The function utilizes the Seurat package for accessing
assay data and performs initial quality control checks and data normalization.
}
