% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_Seurat.R
\name{SDAresChecks}
\alias{SDAresChecks}
\title{Diagnostic Checks on SDA Results}
\usage{
SDAresChecks(results)
}
\value{
Invisible NULL. The function is called for its side effects, which include printing
  diagnostic information and error messages directly to the console.
}
\description{
Executes a series of diagnostic checks on Single-cell Data Analysis (SDA) results to
assess convergence, distribution of loadings and scores, and Posterior Inclusion Probability (PIP)
distributions. The function uses `SDAtools` for generating diagnostic outputs and handles
errors gracefully for each diagnostic step, ensuring that all possible checks are attempted.
}
\details{
This function is designed to be run with SDA results available in the global environment
  or within its scope. It attempts to run convergence checks, loading and score distributions, and
  PIP distributions. If an error occurs during any of the checks, it catches the error, prints
  an error message specific to the failed check, and continues with the next checks.
}
