% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monocle.R
\name{Ser2Monocle_MakeNProcess}
\alias{Ser2Monocle_MakeNProcess}
\title{Ser2Monocle_MakeNProcess}
\usage{
Ser2Monocle_MakeNProcess(
  SeurObj = NULL,
  retunMon = T,
  PCAnDim = 20,
  doUMAP = T,
  min_dist = 0.3,
  n_neighbors = 40,
  dotSNE = F,
  doClust = T,
  ClusLouvRes = 5e-05,
  louvain_iter = 3,
  KeepTopNgenes = 3000,
  minExprGeneDet = 0.1,
  upperScale = 2,
  lowerScale = 2.5,
  assay = NULL
)
}
\arguments{
\item{SeurObj}{A Seurat object with a count matrix and metadata}

\item{retunMon}{Logical value indicating whether to return the Monocle 2 object (default = TRUE)}

\item{PCAnDim}{Number of principal components to retain for dimensionality reduction (default = 20)}

\item{doUMAP}{Logical value indicating whether to perform UMAP reduction (default = TRUE)}

\item{min_dist}{The effective minimum distance between embedded points in the UMAP plot (default = 0.3)}

\item{n_neighbors}{The size of local neighborhood (default = 40)}

\item{dotSNE}{Logical value indicating whether to perform tSNE reduction (default = FALSE)}

\item{doClust}{Logical value indicating whether to perform clustering (default = TRUE)}

\item{ClusLouvRes}{Louvain resolution parameter for clustering (default = 0.00005)}

\item{louvain_iter}{The number of iterations for the Louvain algorithm (default = 3)}

\item{KeepTopNgenes}{The number of top variable genes to use in the analysis (default = 3000)}

\item{minExprGeneDet}{The minimum expression threshold for genes to be considered expressed (default = 0.1)}

\item{upperScale}{The scaling factor for the upper bound for filtering cells by total mRNA expression (default = 2)}

\item{lowerScale}{The scaling factor for the lower bound for filtering cells by total mRNA expression (default = 2.5)}

\item{assay}{The assay type to use (default = NULL)}
}
\value{
A Monocle object
}
\description{
Performs the preprocessing steps for single-cell RNA-seq data using Monocle 2 package
}
\examples{
data("pbmc_small")
pbmc_small <- as.Seurat(pbmc_small)
SeurObj_cds <- Ser2Monocle_MakeNProcess(SeurObj = pbmc_small, retunMon = T,
PCAnDim = 20, doUMAP = T, min_dist=.3,
n_neighbors = 40, dotSNE = F, doClust = T,
ClusLouvRes = 0.00005, louvain_iter = 3,
KeepTopNgenes = 3000, minExprGeneDet = 0.1,
upperScale = 2, lowerScale = 2.5, assay = NULL)
}
