% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_Analysis.R
\name{SetDirectionKEGG}
\alias{SetDirectionKEGG}
\title{SetDirectionKEGG}
\usage{
SetDirectionKEGG(df = NULL)
}
\arguments{
\item{df}{Data frame containing KEGG analysis results.}
}
\value{
Input data frame with two extra columns: Direction (e.g. Up and Down for up and down expressed pathways, respectively), and Selected_PV (e.g. -log10(pval) of the selected up or down p-value)
}
\description{
SetDirectionKEGG
}
