% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{similarity_jaccard_DF}
\alias{similarity_jaccard_DF}
\title{similarity overlap dataframe}
\usage{
similarity_jaccard_DF(df, plot = T)
}
\arguments{
\item{df, input}{dataframe where each column contains N genes/strings}
}
\description{
returns a jaccard similarity matrix comparing similarity overlap of character items
}
