% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_stats.R
\name{similarity_jaccard_LS}
\alias{similarity_jaccard_LS}
\title{similarity_overlap list}
\usage{
similarity_jaccard_LS(VecLs, plot = F)
}
\arguments{
\item{VecLs, }{a list of character vectors}
}
\description{
returns a jaccard similarity matrix comparing similarity overlap of character items
}
