% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{similarity_overlap}
\alias{similarity_overlap}
\title{similarity_overlap}
\usage{
similarity_overlap(VecLs)
}
\arguments{
\item{VecLs, }{a list of character vectors}
}
\description{
returns a jaccard similarity matrix comparing similarity overlap of character items
}
