% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_3D.R
\name{simulate_3D_data}
\alias{simulate_3D_data}
\title{Simulate 3D Data}
\usage{
simulate_3D_data(N, L, P, ncomps)
}
\arguments{
\item{N}{integer, the number of individuals.}

\item{L}{integer, the number of variables.}

\item{P}{integer, the number of replicates.}

\item{ncomps}{integer, the number of components.}
}
\value{
data list, a list containing the generated data including Y, A, B, X, and noise.
}
\description{
The function generates 3D data with N individuals, L variables and P replicates.
}
