% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData_pVL_SIV_ARTv2.R
\name{Simulate_pVL_SIV_ART}
\alias{Simulate_pVL_SIV_ART}
\title{Simulate and reshape plasma viral load (pVL) data for SIV ART study}
\usage{
Simulate_pVL_SIV_ART(
  MonkeyGroups = c("D", "E", "F"),
  PidWindow = 0:748,
  doPlot = T,
  RemoveNAs = F
)
}
\arguments{
\item{MonkeyGroups}{Character vector of monkey group prefixes (e.g., \code{c("D", "E")}). 
These should match the prefixes used in the column names of \code{pVL_clean_imputed}.}

\item{PidWindow}{Post infection range min 0 max 748 days (e.g., \code{0:748}).}

\item{doPlot}{Logical flag indicating whether to generate a plot of the reshaped data.
Defaults to \code{TRUE}.}
}
\value{
A data frame in long format with columns \code{Subject} and \code{ViralLoad},
  suitable for downstream visualization or modeling. Optionally displays a plot.
}
\description{
This function subsets and reshapes plasma viral load data (`pVL_clean_imputed`)
for a given set of monkey groups (e.g., D, E, F) assumed to be columns like D1–D10.
It optionally produces a plot of the reshaped data.
Group D is early anti-PD1, E is late anti-PD1 and F is control for anti-PD1
}
\examples{
Simulate_pVL_SIV_ART(MonkeyGroups = c("D", "F"), doPlot = TRUE)

}
