% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Spatial_based.R
\name{SpatialSer.SlideSeq.local}
\alias{SpatialSer.SlideSeq.local}
\title{Create a Seurat object from Spatial SlideSeq data}
\usage{
SpatialSer.SlideSeq.local(
  dge.path = NULL,
  pos.path = NULL,
  celltype.path = NULL,
  celltype.header = "max_cell_type",
  radius = 2450,
  do.plot = T,
  doFilter = T,
  log_nCount.bool = T
)
}
\arguments{
\item{dge.path}{path to the expression data}

\item{pos.path}{path to the position data}

\item{celltype.path}{path to the cell type data}

\item{celltype.header}{header for the cell type column in celltype.path}

\item{radius}{radius for filtering cells}

\item{do.plot}{logical indicating whether to plot the filtered cells}

\item{doFilter}{logical indicating whether to filter cells}

\item{log_nCount.bool}{logical indicating whether to log transform nCount}
}
\value{
a Seurat object for Spatial SlideSeq data
}
\description{
Create a Seurat object from Spatial SlideSeq data
}
