% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{string_overlap}
\alias{string_overlap}
\title{Calculate gene overlap between two data frames}
\usage{
string_overlap(df1, df2)
}
\arguments{
\item{df1}{A data frame containing genes of interest.}

\item{df2}{A data frame containing genes of interest.}
}
\value{
A list containing three matrices:
\describe{
\item{\code{df1_ol}}{A matrix of gene overlap within \code{df1}.}
\item{\code{df2_ol}}{A matrix of gene overlap within \code{df2}.}
\item{\code{df12_ol}}{A matrix of gene overlap between \code{df1} and \code{df2}.}
}
}
\description{
This function takes in two data frames and calculates the gene overlap within each data frame as well as the overlap between the two data frames.
}
