% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{SubsetSerByCoordinates}
\alias{SubsetSerByCoordinates}
\title{SubsetSerByCoordinates}
\usage{
SubsetSerByCoordinates(
  SerObj,
  reduction_method = "tSNE",
  assay = "RNA",
  x_range = c(-1, 1),
  y_range = c(-1, 1)
)
}
\arguments{
\item{SerObj}{A Seurat Object}

\item{reduction_method}{an input of  = c("tSNE", "UMAP", "PCA") default tSNE}

\item{x_range}{an input of  two numbers defining the x range default c(-1, 1)}

\item{y_range}{aan input of  two numbers defining the y range default c(-1, 1)}
}
\value{
A subset Seurat object.
}
\description{
Remove unwanted HTOS from Seurat objects
}
