% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_themes.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Theme Publication}
\usage{
theme_Publication(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{The base font size for the plot elements.}

\item{base_family}{The base font family for the plot elements.}
}
\description{
This function defines a custom theme for publication-quality plots.
}
