% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_themes.R
\docType{data}
\name{theme_simple}
\alias{theme_simple}
\title{theme_simple remove duplicated extra axes}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 6.
}
\usage{
theme_simple
}
\value{
A ggplot theme
}
\description{
ggplot theme, use as + theme_simple()
}
\keyword{datasets}
