% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_themes.R
\docType{data}
\name{theme_simple2}
\alias{theme_simple2}
\title{theme_simple remove legend, axis, and title with base size 20}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 97.
}
\usage{
theme_simple2
}
\value{
A ggplot theme
}
\description{
ggplot theme, use as + theme_simple()
}
\keyword{datasets}
