% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{UnsupClust_UMAP_proc}
\alias{UnsupClust_UMAP_proc}
\title{Perform unsupervised clustering and UMAP reduction}
\usage{
UnsupClust_UMAP_proc(
  SerObject,
  doClust = T,
  Feats = NULL,
  reduction = NULL,
  dims = NULL,
  resolution = 0.4,
  verbose = F,
  random.seed = 1234,
  doUMAP = T,
  n.components = 2,
  n.neighbors = 60,
  n.epochs = 300,
  min.dist = 0.2,
  spread = 1,
  reduction.key = "UMAPCust_"
)
}
\arguments{
\item{SerObject}{A Seurat obj}

\item{doClust}{Logical if F unsupervised clustering is only done default (T)}

\item{Feats}{A vector of character strings specifying which features to use for clustering. If NULL, all features are used.}

\item{reduction}{The type of dimensionality reduction to apply. If NULL, no reduction is applied. Options are "PCA", "TSNE", "UMAP", or "none".}

\item{dims}{The number of dimensions to use for the reduction. If NULL, the default number of dimensions for the chosen method is used.}

\item{resolution}{The UMAP resolution parameter.}

\item{verbose}{A boolean indicating whether to print progress updates during the clustering process.}

\item{random.seed}{The random seed to use for reproducibility.}

\item{doUMAP}{Logical if F unsupervised clustering is only done default (T)}

\item{n.components}{The number of umap components default 2, or 3 or try 1}

\item{n.neighbors}{The number of neighbors to use for UMAP construction.}

\item{n.epochs}{The number of epochs to use for UMAP construction.}

\item{min.dist}{The minimum distance between UMAP points.}

\item{spread}{The spread parameter for UMAP.}

\item{reduction.key}{A character string specifying the prefix for the output object names.}
}
\value{
seurat obj
}
\description{
This function performs unsupervised clustering on single-cell RNA-seq data and then it does the Uniform Manifold Approximation and Projection (UMAP) dimensionality reduction method. It returns and updated seurat object.
}
