% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{updateClusterLabels}
\alias{updateClusterLabels}
\title{Update Cluster Labels Based on Frequency}
\usage{
updateClusterLabels(SerObj, columnName, sortedColumnName = NULL)
}
\arguments{
\item{SerObj}{A Seurat object containing the data and metadata.}

\item{columnName}{The name of the column in the Seurat object's metadata to analyze
and sort by frequency.}

\item{sortedColumnName}{Optional. The name of the column where the sorted labels
should be stored. If NULL, the function will overwrite the original column.}
}
\value{
The Seurat object with updated cluster labels in its metadata. If
`sortedColumnName` is provided, the updated labels are stored in this new column;
otherwise, the original column is overwritten with the updated labels.
}
\description{
This function updates the labels of a specified column in a Seurat object's metadata
based on the decreasing frequency of occurrence. Optionally, it can store the updated
labels in a new column if a sorted column name is provided; otherwise, it overwrites
the original column.
}
