% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{VlnPlot_pvalue}
\alias{VlnPlot_pvalue}
\title{Generate a violin plot with p-values for a given feature and groupings}
\usage{
VlnPlot_pvalue(
  SerObj,
  feature,
  group.by = "ident",
  slot = "data",
  assay = "RNA",
  colors = col_vector,
  GrpNames = NULL,
  title = "",
  doJitter = T
)
}
\arguments{
\item{SerObj}{A single-cell object, such as a Seurat object.}

\item{feature}{A character string indicating the feature of interest. This should correspond to a column in the slot specified by the slot parameter.}

\item{group.by}{A character string indicating the grouping variable to use for the plot. This should correspond to a column in the object.}

\item{slot}{A character string indicating which slot of the Seurat object to use for the plot. Default is "data".}

\item{assay}{A character string indicating which assay of the slot to use for the plot. Default is "RNA".}

\item{colors}{A vector of colors to use for the plot. By default, the colors will be generated automatically.}

\item{GrpNames}{A vector of group names to use for the plot. If not specified, the group names will be inferred from the group.by column in the object.}

\item{title}{A character string indicating the plot title. Default is no title.}

\item{doJitter}{A logical value indicating whether to add jitter to the plot. Default is TRUE.}
}
\value{
A ggplot object with a violin plot and p-values for each pairwise comparison.
}
\description{
This function takes in a single-cell object, a feature of interest, and grouping variables to generate a violin plot with p-values. The p-values are generated by a Wilcoxon rank sum test comparing the expression levels of the feature of interest between each group. By default, the plot is grouped by the "ident" column in the object, but this can be changed with the group.by parameter.
}
