% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{enrichGOFunction}
\alias{enrichGOFunction}
\title{Enrich GO Function}
\usage{
enrichGOFunction(
  gene_set,
  title,
  title_suffix,
  gene_universe,
  p.adjust.thr = 0.05,
  returnPlot = T,
  qrhub,
  labsize = 3,
  fontsize = 0.25,
  base_size = 18,
  max.overlaps = 20,
  MaxNsig = 30
)
}
\arguments{
\item{gene_set}{A vector of gene symbols for the enrichment analysis.}

\item{title}{The title to be added to the plot title.}

\item{title_suffix}{The suffix to be added to the plot title.}

\item{gene_universe}{The vector of gene symbols as the universe.}
}
\value{
A ggplot object displaying the GO enrichment analysis results.
}
\description{
This function performs Gene Ontology (GO) enrichment analysis and generates a
ggplot object for visualization.
}
