% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_Seurat_based.R
\name{FeatThrTab_boxplot}
\alias{FeatThrTab_boxplot}
\title{Generate a boxplot with jittered points for the percentage of values exceeding a threshold.}
\usage{
FeatThrTab_boxplot(
  SerObj,
  FeatName = NULL,
  CutThresh = NULL,
  MetaDataName = "Population",
  TitleExtra = "",
  col = "#A6CEE3",
  MetaDataName2 = "SubjectId",
  ylab_text = "",
  xlab_text = "",
  base_size = 16,
  my_comparisons = NULL,
  limits = c(0, 110),
  GT = T,
  StatMethod = "wilcox.test",
  StatLab = "p.signif"
)
}
\arguments{
\item{SerObj}{The input data or object.}

\item{FeatName}{The name of the feature variable.}

\item{CutThresh}{The threshold value for categorizing values.}

\item{MetaDataName}{The name of the metadata variable for grouping (default: "Population").}

\item{TitleExtra}{Additional title text for the plot.}

\item{col}{The color for the boxplot and jittered points (default: "#A6CEE3").}

\item{MetaDataName2}{The name of the second metadata variable (default: "SubjectId").}

\item{ylab_text}{The label for the y-axis.}

\item{xlab_text}{The label for the x-axis.}

\item{base_size}{The base font size for the plot (default: 16).}

\item{my_comparisons}{A vector specifying pairwise comparisons for statistical testing (default: NULL).}

\item{limits}{A vector to define the limits of the percent axis (default: c(0, 110)).}

\item{GT}{T/F to select greater than and equal to threshold if T else less than (default: T).}

\item{StatMethod}{passed to stat_compare_means only in t.test and in wilcox.test (default: wilcox.test).}

\item{StatLab}{passed to stat_compare_means  p.signif" (shows the significance levels), "p.format" (shows the formatted p value), (default: p.signif).}
}
\value{
A ggplot object representing the boxplot with jittered points.
}
\description{
Generate a boxplot with jittered points for the percentage of values exceeding a threshold.
}
\examples{
# Example usage:
FeatThrTab_boxplot(SerObj, FeatName = "Feature1", CutThresh = 0.5)


}
