% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_based.R
\name{Find_TCR_MAITS}
\alias{Find_TCR_MAITS}
\title{Identify TRAV1-2 cells and Classical MAITS as TRAV1-2 TRAJ33}
\usage{
Find_TCR_MAITS(SerObj, plot = FALSE)
}
\arguments{
\item{SerObj}{A Seurat object containing the data.}
}
\description{
Identify TRAV1-2 cells and Classical MAITS as TRAV1-2 TRAJ33
}
