% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_tables.R
\name{gg_barplot_2waytable}
\alias{gg_barplot_2waytable}
\title{gg_barplot_table, aka OOSAP::PlotMyTable}
\usage{
gg_barplot_2waytable(
  MyTable,
  Title = "",
  legend.position = "bottom",
  PlotCombo = F,
  xlab = "",
  xtext_angle = 45,
  hjust = 0.9,
  vjust = 0.8,
  base_size = 10,
  returnlist = T,
  col_vector = col_vector,
  theme_b = theme_bw(base_size = base_size)
)
}
\arguments{
\item{MyTable}{A 2way table (base R table())}

\item{Title}{A title else ''}

\item{legend.position}{legend position defautl "bottom"}

\item{PlotCombo}{if T plots % and count plots together}

\item{xlab}{x axis label}

\item{xtext_angle}{x axis label}

\item{hjust}{if angle is 90 set to 1}

\item{vjust}{if angle is 90 set to 0.5}

\item{base_size}{numerical}

\item{returnlist}{logical}

\item{col_vector}{a vector of colors}

\item{theme}{a ggplot theme, default theme_bw(base_size = base_size)}
}
\value{
A ggplot
}
\description{
Takes a table and plots a gg barplot
}
