% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations_general.R
\name{gg_barplot_table}
\alias{gg_barplot_table}
\title{gg_barplot_table}
\usage{
gg_barplot_table(TBL, Title = "", gg_theme = NULL)
}
\arguments{
\item{TBL}{A table (base R table())}

\item{Title}{A title else ''}

\item{gg_theme}{a ggplot theme or NULL}
}
\value{
A ggplot
}
\description{
Takes a table and plots a gg barplot
}
