% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{gg_tSNE_df}
\alias{gg_tSNE_df}
\title{gg_tSNE_df}
\usage{
gg_tSNE_df(
  DF,
  SavePath = NULL,
  Markers = NULL,
  splitby = NULL,
  verbose = T,
  size = 0.2,
  alpha = 0.6,
  Title = "",
  width = 2000,
  height = 2300,
  pointsize = 20,
  ncol = 4,
  limits = c(0, 1),
  gg_theme = NULL,
  printPlot = T,
  vridis_col_opt = "A"
)
}
\arguments{
\item{DF}{A data.frame}

\item{SavePath}{A data.frame}

\item{Markers}{A char. vec. that that all must exist in the data.frame.}

\item{splitby}{a single character which data.frame must contain this value in the cols; the resulting plot is faceted by this.}

\item{verbose}{if T verbose output is printed}

\item{Title}{A title else ''}

\item{gg_theme}{a ggplot theme or NULL}
}
\value{
A ggplot
}
\description{
Takes a data.frame with tSNE1 and tSNE2 precomputed and plots it.
}
