% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{range01}
\alias{range01}
\title{Range01}
\usage{
range01(x, MaxN = NULL, MinN = NULL)
}
\arguments{
\item{x}{a numerical vector}

\item{MaxN}{set Max value; NULL default find max(x)}

\item{MinN}{set Min value; NULL default find min(x)}
}
\value{
A numeric vector
}
\description{
Scales the range of input to be between 0 and 1
}
\keyword{range}
